Modularize calculator assets and styling
