+++ js/calculator.js
@@ -1,133 +1,175 @@
-import {
-  breakdownDialog,
-  breakdownBackdrop,
-  breakdownClose,
-  breakdownSummary,
-  breakdownVariant,
-  breakdownPrice,
-  breakdownMeta,
-  breakdownTotal,
-  breakdownTableBody,
-  breakdownChartGraphic,
-  breakdownLegend,
-  accountingDialog,
-  accountingBackdrop,
-  accountingClose,
-  accountingForm,
-  accountingStudentsInput,
-  accountingClassesInput,
-  accountingHoursInput,
-  accountingError,
-  accountingCancel,
-  controls,
-  desiredIncomeFieldMap,
-  desiredIncomeTitle,
-  desiredIncomeLegend,
-  acceptableIncomeBasisLabel,
-  desiredIncomeLabelSpans,
-  desiredIncomeInfoIcons,
-  tablesContainer,
-  assumptionsList
-} from './dom.js';
-import { getFocusableElements, setBodyScrollLock, ensureInfoIconTooltip } from './ui.js';
-
 export function initCalculator() {
+  const breakdownDialog = document.getElementById('lesson-breakdown-dialog');
+  const breakdownBackdrop = document.getElementById('lesson-breakdown-backdrop');
+  const breakdownClose = document.getElementById('lesson-breakdown-close');
+  const breakdownSummary = document.getElementById('lesson-breakdown-summary');
+  const breakdownVariant = document.getElementById('lesson-breakdown-variant');
+  const breakdownPrice = document.getElementById('lesson-breakdown-price');
+  const breakdownMeta = document.getElementById('lesson-breakdown-meta');
+  const breakdownTotal = document.getElementById('lesson-breakdown-total');
+  const breakdownTableBody = document.getElementById('lesson-breakdown-table-body');
+  const breakdownChartGraphic = document.getElementById('lesson-breakdown-chart-graphic');
+  const breakdownLegend = document.getElementById('lesson-breakdown-legend');
+  const howToLink = document.getElementById('how-to-link');
+  const readmeDialog = document.getElementById('readme-dialog');
+  const readmeBody = document.getElementById('readme-body');
+  const readmeClose = document.getElementById('readme-close');
+  const readmeBackdrop = document.getElementById('readme-backdrop');
+  const accountingDialog = document.getElementById('accounting-dialog');
+  const accountingBackdrop = document.getElementById('accounting-backdrop');
+  const accountingClose = document.getElementById('accounting-close');
+  const accountingForm = document.getElementById('accounting-form');
+  const accountingStudentsInput = document.getElementById('accounting-students');
+  const accountingClassesInput = document.getElementById('accounting-classes');
+  const accountingHoursInput = document.getElementById('accounting-hours');
+  const accountingError = document.getElementById('accounting-error');
+  const accountingCancel = document.getElementById('accounting-cancel');
+  let readmeLoaded = false;
+  let readmeLoading = false;
+  let previouslyFocusedElement = null;
   let breakdownTriggerElement = null;
   let activeBreakdownContext = null;
   let accountingTriggerElement = null;
-
+  
+  const controls = {
+    targetNet: document.getElementById('target-net'),
+    targetNetWeek: document.getElementById('target-net-week'),
+    targetNetMonth: document.getElementById('target-net-month'),
+    targetNetAverageWeek: document.getElementById('target-net-average-week'),
+    targetNetAverageMonth: document.getElementById('target-net-average-month'),
+    desiredIncomeToggleDisplay: document.getElementById('desired-income-toggle-display'),
+    desiredIncomeToggleInterpretation: document.getElementById('desired-income-toggle-interpretation'),
+    acceptableIncomeMin: document.getElementById('acceptable-income-min'),
+    acceptableIncomeMax: document.getElementById('acceptable-income-max'),
+    acceptableIncomeBasisMonthly: document.getElementById('acceptable-income-basis-monthly'),
+    acceptableIncomeBasisAnnual: document.getElementById('acceptable-income-basis-annual'),
+    taxRate: document.getElementById('tax-rate'),
+    fixedCosts: document.getElementById('fixed-costs'),
+    variableCostPerClass: document.getElementById('variable-cost-class'),
+    variableCostPerStudent: document.getElementById('variable-cost-student'),
+    variableCostPerStudentMonthly: document.getElementById('variable-cost-student-monthly'),
+    vatRate: document.getElementById('vat-rate'),
+    classesPerWeek: document.getElementById('classes-per-week'),
+    studentsPerClass: document.getElementById('students-per-class'),
+    hoursPerLesson: document.getElementById('hours-per-lesson'),
+    lessonCost: document.getElementById('lesson-cost'),
+    lessonPriceMin: document.getElementById('lesson-price-min'),
+    lessonPriceMax: document.getElementById('lesson-price-max'),
+    monthsOff: document.getElementById('months-off'),
+    weeksOffCycle: document.getElementById('weeks-off-cycle'),
+    daysOffWeek: document.getElementById('days-off-week'),
+    buffer: document.getElementById('buffer'),
+    currencySymbol: document.getElementById('currency-symbol'),
+    recalcButton: document.getElementById('recalculate'),
+    downloadCsv: document.getElementById('download-csv'),
+    accountingExample: document.getElementById('open-accounting-example'),
+    statusMessage: document.getElementById('status-message'),
+    workingWeeksDisplay: document.getElementById('working-weeks-display'),
+    workingDaysDisplay: document.getElementById('working-days-display'),
+    rememberInputs: document.getElementById('remember-inputs'),
+    resetSavedInputs: document.getElementById('reset-saved-inputs')
+  };
+  
+  const desiredIncomeFieldMap = {
+    year: controls.targetNet,
+    week: controls.targetNetWeek,
+    month: controls.targetNetMonth,
+    avgWeek: controls.targetNetAverageWeek,
+    avgMonth: controls.targetNetAverageMonth
+  };
+  
   const COLLAPSIBLE_SECTION_SELECTOR = '.control-section[data-collapsible]';
-
+  
   function getSectionLabel(section, index = 0) {
     if (!section) {
       return `Section ${index + 1}`;
     }
-
+  
     const title = section.querySelector('.section-title');
     const labelText = title && title.textContent ? title.textContent.trim() : '';
-
+  
     if (labelText) {
       section.dataset.sectionLabel = labelText;
       return labelText;
     }
-
+  
     if (section.dataset.sectionLabel) {
       return section.dataset.sectionLabel;
     }
-
+  
     const fallback = `Section ${index + 1}`;
     section.dataset.sectionLabel = fallback;
     return fallback;
   }
-
+  
   function updateToggleButtonState(toggle, collapsed, sectionLabel) {
     if (!toggle) {
       return;
     }
-
+  
     const expandedText = toggle.getAttribute('data-expanded-text') || 'Collapse section';
     const collapsedText = toggle.getAttribute('data-collapsed-text') || 'Expand section';
     const labelElement = toggle.querySelector('.toggle-label');
     const nextText = collapsed ? collapsedText : expandedText;
-
+  
     if (labelElement) {
       labelElement.textContent = nextText;
     } else {
       toggle.textContent = nextText;
     }
-
+  
     const actionText = collapsed ? collapsedText : expandedText;
     const normalizedAction = actionText.replace(/\s+section$/i, '');
     const normalizedLabel = sectionLabel || 'section';
-
+  
     toggle.setAttribute('aria-expanded', String(!collapsed));
     toggle.setAttribute('aria-label', `${normalizedAction} ${normalizedLabel} section`);
     toggle.setAttribute('title', `${normalizedAction} ${normalizedLabel} section`);
   }
-
+  
   const collapsibleSections = new Map();
-
+  
   function registerCollapsibleSection(section) {
     if (!(section instanceof HTMLElement)) {
       return;
     }
-
+  
     const body = section.querySelector('.section-body');
     if (!(body instanceof HTMLElement)) {
       return;
     }
-
+  
     const existing = collapsibleSections.get(section);
     const listeners = existing ? existing.listeners : new Map();
     const collapsed = section.classList.contains('collapsed');
     const fallbackIndex = collapsibleSections.size;
     const sectionLabel = getSectionLabel(section, fallbackIndex);
     section.dataset.sectionLabel = sectionLabel;
-
+  
     if (!body.id) {
       const baseId = sectionLabel
         .toLowerCase()
         .replace(/[^a-z0-9]+/g, '-')
         .replace(/(^-|-$)/g, '') || `section-${fallbackIndex + 1}`;
-
+  
       let uniqueId = baseId;
       let attempt = 1;
       while (document.getElementById(uniqueId)) {
         attempt += 1;
         uniqueId = `${baseId}-${attempt}`;
       }
-
+  
       body.id = uniqueId;
     }
-
+  
     const seenToggles = new Set();
     section.querySelectorAll('.section-toggle').forEach(toggle => {
       if (!(toggle instanceof HTMLElement)) {
         return;
       }
-
+  
       seenToggles.add(toggle);
-
+  
       if (!listeners.has(toggle)) {
         const handler = event => {
           event.preventDefault();
@@ -136,11 +178,11 @@ export function initCalculator() {
         toggle.addEventListener('click', handler);
         listeners.set(toggle, handler);
       }
-
+  
       toggle.setAttribute('aria-controls', body.id);
       updateToggleButtonState(toggle, collapsed, sectionLabel);
     });
-
+  
     if (existing) {
       for (const [toggle, handler] of listeners) {
         if (!seenToggles.has(toggle)) {
@@ -149,23 +191,23 @@ export function initCalculator() {
         }
       }
     }
-
+  
     collapsibleSections.set(section, { body, listeners });
   }
-
+  
   function unregisterCollapsibleSection(section) {
     const data = collapsibleSections.get(section);
     if (!data) {
       return;
     }
-
+  
     for (const [toggle, handler] of data.listeners) {
       toggle.removeEventListener('click', handler);
     }
-
+  
     collapsibleSections.delete(section);
   }
-
+  
   function captureCollapsibleSectionStates() {
     const states = {};
     document.querySelectorAll(COLLAPSIBLE_SECTION_SELECTOR).forEach(section => {
@@ -181,7 +223,7 @@ export function initCalculator() {
     });
     return states;
   }
-
+  
   function applyCollapsibleSectionStates(states, { skipPersistence = false } = {}) {
     if (!states || typeof states !== 'object') {
       return;
@@ -201,31 +243,31 @@ export function initCalculator() {
       });
     });
   }
-
+  
   function toggleCollapsibleSection(section, { trigger, explicitState, skipPersistence = false } = {}) {
     if (!(section instanceof HTMLElement)) {
       return;
     }
-
+  
     if (!collapsibleSections.has(section)) {
       registerCollapsibleSection(section);
     }
-
+  
     const data = collapsibleSections.get(section);
     if (!data) {
       return;
     }
-
+  
     const isCollapsed = section.classList.contains('collapsed');
     const nextCollapsed = typeof explicitState === 'boolean' ? explicitState : !isCollapsed;
-
+  
     if (nextCollapsed === isCollapsed) {
       const scroll = () => {
         if (typeof section.scrollIntoView === 'function') {
           section.scrollIntoView({ behavior: 'smooth', block: 'start' });
         }
       };
-
+  
       if (typeof requestAnimationFrame === 'function') {
         requestAnimationFrame(scroll);
       } else {
@@ -233,15 +275,15 @@ export function initCalculator() {
       }
       return;
     }
-
+  
     section.classList.toggle('collapsed', nextCollapsed);
     const sectionLabel = getSectionLabel(section, 0);
     section.dataset.sectionLabel = sectionLabel;
-
+  
     for (const toggle of data.listeners.keys()) {
       updateToggleButtonState(toggle, nextCollapsed, sectionLabel);
     }
-
+  
     if (nextCollapsed && trigger && trigger.classList.contains('section-toggle--bottom')) {
       const topToggle = section.querySelector('.section-toggle--top');
       if (topToggle && typeof topToggle.focus === 'function') {
@@ -252,32 +294,32 @@ export function initCalculator() {
         }
       }
     }
-
+  
     const scroll = () => {
       if (typeof section.scrollIntoView === 'function') {
         section.scrollIntoView({ behavior: 'smooth', block: 'start' });
       }
     };
-
+  
     if (typeof requestAnimationFrame === 'function') {
       requestAnimationFrame(scroll);
     } else {
       scroll();
     }
-
+  
     if (!skipPersistence && persistenceEnabled) {
       savePersistedInputs();
     }
   }
-
+  
   function refreshCollapsibleSections(root = document) {
     Array.from(root.querySelectorAll(COLLAPSIBLE_SECTION_SELECTOR)).forEach(section => {
       registerCollapsibleSection(section);
     });
   }
-
+  
   refreshCollapsibleSections(document);
-
+  
   if (typeof MutationObserver === 'function') {
     const observerTarget = document.body || document.documentElement;
     if (observerTarget) {
@@ -287,27 +329,27 @@ export function initCalculator() {
             if (!(node instanceof HTMLElement)) {
               return;
             }
-
+  
             if (node.matches && node.matches(COLLAPSIBLE_SECTION_SELECTOR)) {
               registerCollapsibleSection(node);
             }
-
+  
             if (typeof node.querySelectorAll === 'function') {
               node.querySelectorAll(COLLAPSIBLE_SECTION_SELECTOR).forEach(childSection => {
                 registerCollapsibleSection(childSection);
               });
             }
           });
-
+  
           mutation.removedNodes?.forEach(node => {
             if (!(node instanceof HTMLElement)) {
               return;
             }
-
+  
             if (collapsibleSections.has(node)) {
               unregisterCollapsibleSection(node);
             }
-
+  
             if (typeof node.querySelectorAll === 'function') {
               node.querySelectorAll(COLLAPSIBLE_SECTION_SELECTOR).forEach(childSection => {
                 unregisterCollapsibleSection(childSection);
@@ -316,27 +358,204 @@ export function initCalculator() {
           });
         }
       });
-
+  
       collapsibleObserver.observe(observerTarget, { childList: true, subtree: true });
     }
   }
-
+  
+  function getFocusableElements(container) {
+    if (!container) {
+      return [];
+    }
+  
+    const selectors = [
+      'a[href]',
+      'button:not([disabled])',
+      'textarea:not([disabled])',
+      'input:not([disabled])',
+      'select:not([disabled])',
+      '[tabindex]:not([tabindex="-1"])'
+    ].join(',');
+  
+    return Array.from(container.querySelectorAll(selectors)).filter(element => {
+      if (element.hasAttribute('hidden') || element.getAttribute('aria-hidden') === 'true') {
+        return false;
+      }
+      const rect = element.getBoundingClientRect();
+      return rect.width > 0 && rect.height > 0;
+    });
+  }
+  
+  const activeDialogLocks = new Set();
+  
+  function setBodyScrollLock(source, locked) {
+    if (!source) {
+      return;
+    }
+  
+    if (locked) {
+      activeDialogLocks.add(source);
+    } else {
+      activeDialogLocks.delete(source);
+    }
+  
+    if (activeDialogLocks.size > 0) {
+      document.body.classList.add('dialog-open');
+    } else {
+      document.body.classList.remove('dialog-open');
+    }
+  }
+  
+  function parseMarkdown(text) {
+    if (window.marked && typeof window.marked.parse === 'function') {
+      return window.marked.parse(text);
+    }
+    const escaped = text
+      .replace(/&/g, '&amp;')
+      .replace(/</g, '&lt;')
+      .replace(/>/g, '&gt;');
+    return `<pre>${escaped}</pre>`;
+  }
+  
+  const README_SOURCES = [
+    new URL('README.md', document.baseURI).toString(),
+    'https://raw.githubusercontent.com/jd-d/course-pricing-calculator/main/README.md'
+  ];
+  
+  function fetchReadmeFromSources(sources, index = 0) {
+    if (!Array.isArray(sources) || index >= sources.length) {
+      return Promise.reject(new Error('No README sources available'));
+    }
+  
+    const source = sources[index];
+  
+    return fetch(source)
+      .then(response => {
+        if (!response || !response.ok) {
+          throw new Error(`Failed to load README from ${source} (status ${response ? response.status : 'unknown'})`);
+        }
+        return response.text();
+      })
+      .catch(error => {
+        if (index < sources.length - 1) {
+          if (console && typeof console.warn === 'function') {
+            console.warn(error);
+          }
+  
+          if (readmeBody) {
+            readmeBody.innerHTML = '<p class="readme-status">Retrying with a backup instructions source…</p>';
+          }
+  
+          return fetchReadmeFromSources(sources, index + 1);
+        }
+  
+        throw error;
+      });
+  }
+  
+  function loadReadme() {
+    if (readmeLoaded || readmeLoading || !readmeBody) {
+      return;
+    }
+  
+    readmeLoading = true;
+  
+    fetchReadmeFromSources(README_SOURCES)
+      .then(text => {
+        readmeBody.innerHTML = parseMarkdown(text);
+        readmeLoaded = true;
+      })
+      .catch(() => {
+        readmeBody.innerHTML = '<p class="readme-status readme-status--error">Unable to load instructions. Please check the README file directly.</p>';
+      })
+      .finally(() => {
+        readmeLoading = false;
+      });
+  }
+  
+  function openReadme(event) {
+    if (event) {
+      event.preventDefault();
+    }
+  
+    if (!readmeDialog || !readmeBody) {
+      return;
+    }
+  
+    previouslyFocusedElement = document.activeElement instanceof HTMLElement ? document.activeElement : null;
+    readmeDialog.hidden = false;
+    setBodyScrollLock('readme', true);
+  
+    if (!readmeLoaded) {
+      readmeBody.innerHTML = '<p class="readme-status">Loading instructions…</p>';
+      loadReadme();
+    }
+  
+    window.requestAnimationFrame(() => {
+      if (readmeClose) {
+        readmeClose.focus();
+      }
+    });
+  
+    document.addEventListener('keydown', handleReadmeKeydown);
+  }
+  
+  function closeReadme() {
+    if (!readmeDialog) {
+      return;
+    }
+  
+    readmeDialog.hidden = true;
+    setBodyScrollLock('readme', false);
+    document.removeEventListener('keydown', handleReadmeKeydown);
+  
+    if (previouslyFocusedElement && typeof previouslyFocusedElement.focus === 'function') {
+      previouslyFocusedElement.focus();
+    }
+  }
+  
+  function handleReadmeKeydown(event) {
+    if (event.key === 'Escape') {
+      closeReadme();
+      return;
+    }
+  
+    if (event.key === 'Tab' && readmeDialog && !readmeDialog.hidden) {
+      const focusable = getFocusableElements(readmeDialog);
+      if (focusable.length === 0) {
+        event.preventDefault();
+        return;
+      }
+  
+      const first = focusable[0];
+      const last = focusable[focusable.length - 1];
+  
+      if (!event.shiftKey && document.activeElement === last) {
+        event.preventDefault();
+        first.focus();
+      } else if (event.shiftKey && document.activeElement === first) {
+        event.preventDefault();
+        last.focus();
+      }
+    }
+  }
+  
   function openAccountingDialog(event) {
     if (event) {
       event.preventDefault();
     }
-
+  
     if (!accountingDialog) {
       return;
     }
-
+  
     if (!latestInputsSnapshot) {
       const snapshotInputs = getInputs();
       latestInputsSnapshot = cloneInputs(snapshotInputs);
     }
-
+  
     accountingTriggerElement = document.activeElement instanceof HTMLElement ? document.activeElement : null;
-
+  
     const snapshot = latestInputsSnapshot || {};
     const setFieldValue = (input, value, digits = 2) => {
       if (!(input instanceof HTMLInputElement)) {
@@ -348,7 +567,7 @@ export function initCalculator() {
         input.value = '';
       }
     };
-
+  
     const defaultStudents = Array.isArray(snapshot.studentsPerClass) && snapshot.studentsPerClass.length
       ? snapshot.studentsPerClass[0]
       : null;
@@ -356,64 +575,64 @@ export function initCalculator() {
       ? snapshot.classesPerWeek[0]
       : null;
     const defaultHours = Number.isFinite(snapshot.hoursPerLesson) ? snapshot.hoursPerLesson : null;
-
+  
     setFieldValue(accountingStudentsInput, defaultStudents, 0);
     setFieldValue(accountingClassesInput, defaultClasses, 2);
     setFieldValue(accountingHoursInput, defaultHours, 2);
-
+  
     if (accountingError) {
       accountingError.textContent = '';
     }
-
+  
     accountingDialog.hidden = false;
     setBodyScrollLock('accounting', true);
-
+  
     window.requestAnimationFrame(() => {
       if (accountingStudentsInput instanceof HTMLInputElement) {
         accountingStudentsInput.focus();
         accountingStudentsInput.select();
       }
     });
-
+  
     document.addEventListener('keydown', handleAccountingKeydown);
   }
-
+  
   function closeAccountingDialog() {
     if (!accountingDialog) {
       return;
     }
-
+  
     accountingDialog.hidden = true;
     setBodyScrollLock('accounting', false);
     document.removeEventListener('keydown', handleAccountingKeydown);
-
+  
     if (accountingError) {
       accountingError.textContent = '';
     }
-
+  
     const trigger = accountingTriggerElement;
     accountingTriggerElement = null;
     if (trigger && typeof trigger.focus === 'function') {
       trigger.focus();
     }
   }
-
+  
   function handleAccountingKeydown(event) {
     if (event.key === 'Escape') {
       closeAccountingDialog();
       return;
     }
-
+  
     if (event.key === 'Tab' && accountingDialog && !accountingDialog.hidden) {
       const focusable = getFocusableElements(accountingDialog);
       if (focusable.length === 0) {
         event.preventDefault();
         return;
       }
-
+  
       const first = focusable[0];
       const last = focusable[focusable.length - 1];
-
+  
       if (!event.shiftKey && document.activeElement === last) {
         event.preventDefault();
         first.focus();
@@ -423,10 +642,10 @@ export function initCalculator() {
       }
     }
   }
-
+  
   function handleAccountingSubmit(event) {
     event.preventDefault();
-
+  
     const studentsValue = Number(accountingStudentsInput?.value);
     if (!Number.isFinite(studentsValue) || studentsValue <= 0) {
       if (accountingError) {
@@ -438,7 +657,7 @@ export function initCalculator() {
       }
       return;
     }
-
+  
     const classesPerWeekValue = Number(accountingClassesInput?.value);
     if (!Number.isFinite(classesPerWeekValue) || classesPerWeekValue <= 0) {
       if (accountingError) {
@@ -450,7 +669,7 @@ export function initCalculator() {
       }
       return;
     }
-
+  
     const hoursPerClassValue = Number(accountingHoursInput?.value);
     if (!Number.isFinite(hoursPerClassValue) || hoursPerClassValue <= 0) {
       if (accountingError) {
@@ -462,18 +681,18 @@ export function initCalculator() {
       }
       return;
     }
-
+  
     if (!Array.isArray(latestPricingData) || !latestPricingData.length) {
       if (accountingError) {
         accountingError.textContent = 'Generate the pricing table first by adding class and student values, then press Recalculate.';
       }
       return;
     }
-
+  
     if (!latestInputsSnapshot) {
       latestInputsSnapshot = cloneInputs(getInputs());
     }
-
+  
     const combination = findBestPricingCombination(studentsValue, classesPerWeekValue, latestPricingData);
     if (!combination) {
       if (accountingError) {
@@ -481,7 +700,7 @@ export function initCalculator() {
       }
       return;
     }
-
+  
     const reportHtml = buildAccountingReport({
       studentsRequested: studentsValue,
       classesPerWeekRequested: classesPerWeekValue,
@@ -495,14 +714,14 @@ export function initCalculator() {
         ? combination.column?.manualNet?.bufferedMonthly
         : combination.column?.buffered?.monthlyNet
     });
-
+  
     if (!reportHtml) {
       if (accountingError) {
         accountingError.textContent = 'Unable to assemble the report. Please recalculate the table and try again.';
       }
       return;
     }
-
+  
     const blob = new Blob([reportHtml], { type: 'text/html;charset=utf-8;' });
     const url = URL.createObjectURL(blob);
     const link = document.createElement('a');
@@ -513,13 +732,13 @@ export function initCalculator() {
     link.click();
     document.body.removeChild(link);
     URL.revokeObjectURL(url);
-
+  
     if (accountingError) {
       accountingError.textContent = '';
     }
-
+  
     closeAccountingDialog();
-
+  
     if (controls.statusMessage) {
       controls.statusMessage.textContent = 'HTML download started. File contains an accountant-style monthly breakdown.';
       window.setTimeout(() => {
@@ -529,28 +748,43 @@ export function initCalculator() {
       }, 2500);
     }
   }
-
+  
+  if (howToLink) {
+    howToLink.addEventListener('click', openReadme);
+  }
+  
+  if (readmeClose) {
+    readmeClose.addEventListener('click', closeReadme);
+  }
+  
+  if (readmeBackdrop) {
+    readmeBackdrop.addEventListener('click', closeReadme);
+  }
+  
+  if (readmeDialog) {
+    readmeDialog.addEventListener('click', event => {
+      if (event.target === readmeDialog) {
         closeReadme();
       }
     });
   }
-
+  
   if (controls.accountingExample instanceof HTMLButtonElement) {
     controls.accountingExample.addEventListener('click', openAccountingDialog);
   }
-
+  
   if (accountingClose) {
     accountingClose.addEventListener('click', closeAccountingDialog);
   }
-
+  
   if (accountingCancel) {
     accountingCancel.addEventListener('click', closeAccountingDialog);
   }
-
+  
   if (accountingBackdrop) {
     accountingBackdrop.addEventListener('click', closeAccountingDialog);
   }
-
+  
   if (accountingDialog) {
     accountingDialog.addEventListener('click', event => {
       if (event.target === accountingDialog) {
@@ -558,11 +792,11 @@ export function initCalculator() {
       }
     });
   }
-
+  
   if (accountingForm) {
     accountingForm.addEventListener('submit', handleAccountingSubmit);
   }
-
+  
   [accountingStudentsInput, accountingClassesInput, accountingHoursInput].forEach(input => {
     if (input instanceof HTMLInputElement) {
       input.addEventListener('input', () => {
@@ -572,15 +806,15 @@ export function initCalculator() {
       });
     }
   });
-
+  
   if (breakdownClose) {
     breakdownClose.addEventListener('click', closeBreakdownDialog);
   }
-
+  
   if (breakdownBackdrop) {
     breakdownBackdrop.addEventListener('click', closeBreakdownDialog);
   }
-
+  
   if (breakdownDialog) {
     breakdownDialog.addEventListener('click', event => {
       if (event.target === breakdownDialog) {
@@ -588,9 +822,81 @@ export function initCalculator() {
       }
     });
   }
-
+  
+  const themeToggleButton = document.getElementById('theme-toggle');
+  const themeToggleLabel = themeToggleButton ? themeToggleButton.querySelector('.toggle-label') : null;
+  const THEME_STORAGE_KEY = 'course-pricing-theme';
+  const prefersDarkScheme = typeof window.matchMedia === 'function'
+    ? window.matchMedia('(prefers-color-scheme: dark)')
+    : { matches: true };
+  
+  function getStoredTheme() {
+    try {
+      const stored = localStorage.getItem(THEME_STORAGE_KEY);
+      if (stored === 'light' || stored === 'dark') {
+        return stored;
+      }
+    } catch (error) {
+      // Ignore storage access errors (e.g. private browsing restrictions)
+    }
+    return null;
+  }
+  
+  function applyThemePreference(theme, { save = true } = {}) {
+    const normalized = theme === 'light' ? 'light' : 'dark';
+    document.body.dataset.theme = normalized;
+  
+    if (save) {
+      try {
+        localStorage.setItem(THEME_STORAGE_KEY, normalized);
+      } catch (error) {
+        // Ignore storage access errors
+      }
+    }
+  
+    if (themeToggleButton && themeToggleLabel) {
+      const isDark = normalized === 'dark';
+      const actionLabel = `Switch to ${isDark ? 'light' : 'dark'} mode`;
+      themeToggleButton.setAttribute('aria-pressed', String(isDark));
+      themeToggleButton.setAttribute('aria-label', actionLabel);
+      themeToggleButton.setAttribute('title', actionLabel);
+      themeToggleLabel.textContent = isDark ? 'Dark mode' : 'Light mode';
+    }
+  
+    document.dispatchEvent(
+      new CustomEvent('themechange', { detail: normalized })
+    );
+  }
+  
+  const storedTheme = getStoredTheme();
+  let respectSystemPreference = !storedTheme;
+  
+  applyThemePreference(storedTheme || (prefersDarkScheme.matches ? 'dark' : 'light'), {
+    save: Boolean(storedTheme)
+  });
+  
+  function handlePreferenceChange(event) {
+    if (respectSystemPreference) {
+      applyThemePreference(event.matches ? 'dark' : 'light', { save: false });
+    }
+  }
+  
+  if (typeof prefersDarkScheme.addEventListener === 'function') {
+    prefersDarkScheme.addEventListener('change', handlePreferenceChange);
+  } else if (typeof prefersDarkScheme.addListener === 'function') {
+    prefersDarkScheme.addListener(handlePreferenceChange);
+  }
+  
+  if (themeToggleButton) {
+    themeToggleButton.addEventListener('click', () => {
+      const nextTheme = document.body.dataset.theme === 'dark' ? 'light' : 'dark';
+      respectSystemPreference = false;
+      applyThemePreference(nextTheme);
+    });
+  }
+  
   const numberFormatter = new Intl.NumberFormat(undefined, { maximumFractionDigits: 0 });
-
+  
   let latestResults = [];
   let showBasePrices = false;
   let targetNetBasis = 'year';
@@ -625,12 +931,12 @@ export function initCalculator() {
     'target-net-average-week': 'avgWeek',
     'target-net-average-month': 'avgMonth'
   };
-
+  
   function getDesiredIncomeField(key) {
     const field = desiredIncomeFieldMap[key];
     return field instanceof HTMLInputElement ? field : null;
   }
-
+  
   function readDesiredIncomeNet(key, fallback = null) {
     const field = getDesiredIncomeField(key);
     if (!field) {
@@ -652,7 +958,7 @@ export function initCalculator() {
     }
     return fallback;
   }
-
+  
   function writeDesiredIncomeNet(key, value) {
     const field = getDesiredIncomeField(key);
     if (!field) {
@@ -664,7 +970,7 @@ export function initCalculator() {
       delete field.dataset.netValue;
     }
   }
-
+  
   function captureDesiredIncomeNetValues() {
     return Object.keys(desiredIncomeFieldMap).reduce((accumulator, key) => {
       const value = readDesiredIncomeNet(key, null);
@@ -674,7 +980,7 @@ export function initCalculator() {
       return accumulator;
     }, {});
   }
-
+  
   function restoreDesiredIncomeNetValues(values) {
     if (!values || typeof values !== 'object') {
       return;
@@ -685,7 +991,7 @@ export function initCalculator() {
       }
     });
   }
-
+  
   function convertGrossToNet(value, taxRate) {
     if (!Number.isFinite(value)) {
       return null;
@@ -693,7 +999,7 @@ export function initCalculator() {
     const normalizedRate = Math.min(Math.max(taxRate, 0), 0.9999);
     return value * (1 - normalizedRate);
   }
-
+  
   function convertNetToGross(value, taxRate) {
     if (!Number.isFinite(value)) {
       return null;
@@ -702,7 +1008,7 @@ export function initCalculator() {
     const denominator = Math.max(1 - normalizedRate, 0.0001);
     return value / denominator;
   }
-
+  
   function setDesiredIncomeLockedGrossValue(key, value) {
     if (!desiredIncomeLockedAsGross) {
       return;
@@ -721,7 +1027,7 @@ export function initCalculator() {
     }
     desiredIncomeLockedGrossValues[key] = Math.max(value, 0);
   }
-
+  
   function getDesiredIncomeLockedGrossValue(key, fallback = null) {
     if (!desiredIncomeLockedGrossValues) {
       return fallback;
@@ -729,11 +1035,11 @@ export function initCalculator() {
     const stored = desiredIncomeLockedGrossValues[key];
     return Number.isFinite(stored) ? stored : fallback;
   }
-
+  
   function clearDesiredIncomeLockedGrossStore() {
     desiredIncomeLockedGrossValues = null;
   }
-
+  
   function setLockedAcceptableIncomeGross(type, value, basis = acceptableIncomeBasis) {
     if (!desiredIncomeLockedAsGross) {
       return;
@@ -758,7 +1064,7 @@ export function initCalculator() {
       acceptableIncomeLockedGrossMaxBasis = normalizedBasis;
     }
   }
-
+  
   function getLockedAcceptableIncomeGross(type) {
     const value = type === 'min' ? acceptableIncomeLockedGrossMin : acceptableIncomeLockedGrossMax;
     const basis = type === 'min' ? acceptableIncomeLockedGrossMinBasis : acceptableIncomeLockedGrossMaxBasis;
@@ -767,27 +1073,27 @@ export function initCalculator() {
       basis: ACCEPTABLE_INCOME_BASIS_VALUES.includes(basis) ? basis : null
     };
   }
-
+  
   function clearLockedAcceptableIncomeGross() {
     acceptableIncomeLockedGrossMin = null;
     acceptableIncomeLockedGrossMax = null;
     acceptableIncomeLockedGrossMinBasis = null;
     acceptableIncomeLockedGrossMaxBasis = null;
   }
-
+  
   function synchronizeLockedDesiredIncomeNetValues(taxRate) {
     if (!desiredIncomeLockedAsGross) {
       return;
     }
-
+  
     Object.entries(desiredIncomeFieldMap).forEach(([key, field]) => {
       if (!(field instanceof HTMLInputElement)) {
         return;
       }
-
+  
       const cached = getDesiredIncomeLockedGrossValue(key, null);
       let grossValue = Number.isFinite(cached) ? cached : null;
-
+  
       const raw = typeof field.value === 'string' ? field.value.trim() : '';
       if (raw !== '') {
         const parsed = Number(raw);
@@ -802,30 +1108,30 @@ export function initCalculator() {
           }
         }
       }
-
+  
       if (!Number.isFinite(grossValue)) {
         writeDesiredIncomeNet(key, null);
         setDesiredIncomeLockedGrossValue(key, null);
         return;
       }
-
+  
       const netValue = convertGrossToNet(grossValue, taxRate);
       if (!Number.isFinite(netValue)) {
         return;
       }
-
+  
       writeDesiredIncomeNet(key, Math.max(netValue, 0));
       setDesiredIncomeLockedGrossValue(key, grossValue);
     });
   }
-
+  
   function synchronizeLockedAcceptableIncomeNetValues(taxRate, activeMonths) {
     if (!desiredIncomeLockedAsGross) {
       return;
     }
-
+  
     const months = Number.isFinite(activeMonths) && activeMonths > 0 ? activeMonths : MONTHS_PER_YEAR;
-
+  
     const convertGrossBasisToAnnualNet = (grossValue, basis) => {
       if (!Number.isFinite(grossValue)) {
         return null;
@@ -842,7 +1148,7 @@ export function initCalculator() {
       }
       return netBasis * months;
     };
-
+  
     const { value: minGross, basis: minBasis } = getLockedAcceptableIncomeGross('min');
     if (Number.isFinite(minGross)) {
       const annualNet = convertGrossBasisToAnnualNet(minGross, minBasis);
@@ -850,7 +1156,7 @@ export function initCalculator() {
     } else {
       acceptableIncomeMinAnnualNet = null;
     }
-
+  
     const { value: maxGross, basis: maxBasis } = getLockedAcceptableIncomeGross('max');
     if (Number.isFinite(maxGross)) {
       const annualNet = convertGrossBasisToAnnualNet(maxGross, maxBasis);
@@ -859,12 +1165,12 @@ export function initCalculator() {
       acceptableIncomeMaxAnnualNet = null;
     }
   }
-
+  
   const WEEKS_PER_YEAR = 52;
   const MONTHS_PER_YEAR = 12;
   const BASE_WORK_DAYS_PER_WEEK = 7;
   const TARGET_NET_DEFAULT = 50000;
-
+  
   const ACCEPTABLE_INCOME_BASIS_VALUES = ['monthly', 'annual'];
   let acceptableIncomeBasis = 'monthly';
   let acceptableIncomeMinAnnualNet = null;
@@ -872,7 +1178,7 @@ export function initCalculator() {
   let previousAcceptableIncome = null;
   let lastActiveMonths = MONTHS_PER_YEAR;
   let lastWorkingWeeks = WEEKS_PER_YEAR;
-
+  
   function updateDesiredIncomeTitle() {
     const isGross = desiredIncomeDisplayMode === 'gross';
     const title = isGross ? 'Desired Gross Income' : 'Desired Net Income';
@@ -883,7 +1189,7 @@ export function initCalculator() {
       desiredIncomeLegend.textContent = isGross ? 'Desired gross income' : 'Desired net income';
     }
   }
-
+  
   function updateDesiredIncomeLabels() {
     const isGross = desiredIncomeDisplayMode === 'gross';
     desiredIncomeLabelSpans.forEach(span => {
@@ -893,7 +1199,7 @@ export function initCalculator() {
       }
     });
   }
-
+  
   function updateDesiredIncomeTooltips() {
     const isGross = desiredIncomeDisplayMode === 'gross';
     desiredIncomeInfoIcons.forEach(icon => {
@@ -912,7 +1218,7 @@ export function initCalculator() {
       }
     });
   }
-
+  
   function updateAcceptableIncomeBasisLabel() {
     if (!acceptableIncomeBasisLabel) {
       return;
@@ -921,7 +1227,7 @@ export function initCalculator() {
     const basisLabel = acceptableIncomeBasis === 'annual' ? 'Annual' : 'Monthly';
     acceptableIncomeBasisLabel.textContent = `${basisLabel} (${modeLabel})`;
   }
-
+  
   function clearDesiredIncomeEditingState() {
     Object.values(desiredIncomeFieldMap).forEach(field => {
       if (field instanceof HTMLInputElement) {
@@ -929,7 +1235,7 @@ export function initCalculator() {
       }
     });
   }
-
+  
   function clearAcceptableIncomeEditingState() {
     [controls.acceptableIncomeMin, controls.acceptableIncomeMax].forEach(field => {
       if (field instanceof HTMLInputElement) {
@@ -937,14 +1243,14 @@ export function initCalculator() {
       }
     });
   }
-
+  
   function refreshDesiredIncomeDisplay(derivedNetValues, taxRate, options = {}) {
     const forceUpdate = options.force === true;
     updateDesiredIncomeTitle();
     updateDesiredIncomeLabels();
     updateAcceptableIncomeBasisLabel();
     updateDesiredIncomeTooltips();
-
+  
     Object.entries(desiredIncomeFieldMap).forEach(([key, field]) => {
       if (!(field instanceof HTMLInputElement)) {
         return;
@@ -966,28 +1272,28 @@ export function initCalculator() {
         netValue = null;
         writeDesiredIncomeNet(key, null);
       }
-
+  
       const isEditing = field.dataset.editing === 'true';
-
+  
       if (!Number.isFinite(netValue)) {
         if (!isEditing || forceUpdate) {
           field.value = '';
         }
         return;
       }
-
+  
       const displayValue = desiredIncomeDisplayMode === 'gross'
         ? convertNetToGross(netValue, taxRate)
         : netValue;
-
+  
       if (isEditing && !forceUpdate) {
         return;
       }
-
+  
       field.value = formatFixed(displayValue, 2);
     });
   }
-
+  
   function refreshAcceptableIncomeDisplay(taxRate, options = {}) {
     const forceUpdate = options.force === true;
     updateAcceptableIncomeBasisLabel();
@@ -998,7 +1304,7 @@ export function initCalculator() {
       ? controls.acceptableIncomeMax
       : null;
     const monthsForRange = lastActiveMonths > 0 ? lastActiveMonths : MONTHS_PER_YEAR;
-
+  
     const toDisplay = annualNet => {
       if (!Number.isFinite(annualNet)) {
         return '';
@@ -1011,7 +1317,7 @@ export function initCalculator() {
         : netBasis;
       return Number.isFinite(value) ? formatFixed(value, 2) : '';
     };
-
+  
     if (minInput) {
       if (minInput.dataset.editing === 'true' && !forceUpdate) {
         // Keep the user's in-progress value.
@@ -1027,7 +1333,7 @@ export function initCalculator() {
       }
     }
   }
-
+  
   function initializeDesiredIncomeDatasets() {
     Object.entries(desiredIncomeFieldMap).forEach(([key, field]) => {
       if (!(field instanceof HTMLInputElement)) {
@@ -1044,7 +1350,7 @@ export function initCalculator() {
       }
     });
   }
-
+  
   function updateDesiredIncomeFromField(basis) {
     const field = getDesiredIncomeField(basis);
     if (!(field instanceof HTMLInputElement)) {
@@ -1085,7 +1391,7 @@ export function initCalculator() {
       delete desiredIncomeLockedGrossValues[basis];
     }
   }
-
+  
   function updateAcceptableIncomeFromInput(type) {
     const input = type === 'min' ? controls.acceptableIncomeMin : controls.acceptableIncomeMax;
     if (!(input instanceof HTMLInputElement)) {
@@ -1136,7 +1442,17 @@ export function initCalculator() {
     }
     previousAcceptableIncome = null;
   }
-
+  
+  const desiredIncomeTitle = document.getElementById('desired-income-title');
+  const desiredIncomeLegend = document.getElementById('section-desired-income-legend');
+  const acceptableIncomeBasisLabel = document.getElementById('acceptable-income-current-basis');
+  const desiredIncomeLabelSpans = Array.from(
+    document.querySelectorAll('#section-desired-income .label-text[data-net-label][data-gross-label]')
+  );
+  const desiredIncomeInfoIcons = Array.from(
+    document.querySelectorAll('#section-desired-income .info-icon[data-net-tooltip][data-gross-tooltip]')
+  );
+  
   const fixedCostFields = {
     location: {
       monthly: document.getElementById('fixed-cost-location-monthly'),
@@ -1175,21 +1491,21 @@ export function initCalculator() {
       annual: document.getElementById('fixed-cost-development-annual')
     }
   };
-
+  
   controls.fixedCostFields = fixedCostFields;
-
+  
   const rememberInputsToggle = controls.rememberInputs instanceof HTMLInputElement ? controls.rememberInputs : null;
   const resetSavedInputsButton = controls.resetSavedInputs instanceof HTMLButtonElement ? controls.resetSavedInputs : null;
   const PERSISTENCE_ENABLED_KEY = 'course-pricing-save-enabled';
   const PERSISTENCE_VALUES_KEY = 'course-pricing-saved-inputs';
   let persistenceEnabled = false;
   let persistableInputsCache = null;
-
+  
   initializeDesiredIncomeDatasets();
   updateDesiredIncomeTitle();
   updateDesiredIncomeLabels();
   updateAcceptableIncomeBasisLabel();
-
+  
   function getPersistableInputs() {
     if (persistableInputsCache) {
       return persistableInputsCache;
@@ -1204,7 +1520,7 @@ export function initCalculator() {
     });
     return persistableInputsCache;
   }
-
+  
   function captureInputValues(inputs = getPersistableInputs()) {
     const values = inputs.reduce((accumulator, input) => {
       if (!(input instanceof HTMLInputElement) || !input.id) {
@@ -1234,9 +1550,9 @@ export function initCalculator() {
     });
     return values;
   }
-
+  
   const defaultInputValues = captureInputValues();
-
+  
   function readPersistenceEnabled() {
     try {
       return localStorage.getItem(PERSISTENCE_ENABLED_KEY) === 'true';
@@ -1244,7 +1560,7 @@ export function initCalculator() {
       return false;
     }
   }
-
+  
   function readPersistedValues() {
     try {
       const raw = localStorage.getItem(PERSISTENCE_VALUES_KEY);
@@ -1260,7 +1576,7 @@ export function initCalculator() {
     }
     return null;
   }
-
+  
   function clearPersistedInputs() {
     try {
       localStorage.removeItem(PERSISTENCE_VALUES_KEY);
@@ -1269,7 +1585,7 @@ export function initCalculator() {
       // Ignore storage errors
     }
   }
-
+  
   function applyInputValues(values, options = {}) {
     const { skipRender = false } = options;
     if (!values || typeof values !== 'object') {
@@ -1368,7 +1684,7 @@ export function initCalculator() {
       render();
     }
   }
-
+  
   function savePersistedInputs() {
     if (!persistenceEnabled) {
       return;
@@ -1384,7 +1700,7 @@ export function initCalculator() {
       }
     }
   }
-
+  
   function handlePersistableInputMutation(event) {
     if (event && event.target instanceof HTMLInputElement) {
       const nextBasis = TARGET_NET_BASIS_BY_INPUT_ID[event.target.id];
@@ -1396,7 +1712,7 @@ export function initCalculator() {
       savePersistedInputs();
     }
   }
-
+  
   function initializePersistence() {
     if (rememberInputsToggle) {
       rememberInputsToggle.checked = readPersistenceEnabled();
@@ -1404,17 +1720,17 @@ export function initCalculator() {
     } else {
       persistenceEnabled = false;
     }
-
+  
     const storedValues = persistenceEnabled ? readPersistedValues() : null;
     if (storedValues) {
       applyInputValues(storedValues, { skipRender: true });
     }
-
+  
     getPersistableInputs().forEach(input => {
       input.addEventListener('input', handlePersistableInputMutation);
       input.addEventListener('change', handlePersistableInputMutation);
     });
-
+  
     if (rememberInputsToggle) {
       rememberInputsToggle.addEventListener('change', () => {
         persistenceEnabled = rememberInputsToggle.checked;
@@ -1425,7 +1741,7 @@ export function initCalculator() {
         }
       });
     }
-
+  
     if (resetSavedInputsButton) {
       resetSavedInputsButton.addEventListener('click', () => {
         applyInputValues(defaultInputValues);
@@ -1437,16 +1753,201 @@ export function initCalculator() {
       });
     }
   }
-
+  
+  const tablesContainer = document.getElementById('tables-container');
   let tablesLayoutUpdateScheduled = false;
-
+  const assumptionsList = document.getElementById('assumptions-list');
+  
+  const INFO_ICON_ACTIVE_CLASS = 'is-active';
+  const TOOLTIP_VIEWPORT_PADDING = 16;
+  let activeInfoIcon = null;
+  let infoIconTooltipId = 0;
+  
+  function isInfoIcon(element) {
+    return element instanceof HTMLElement && element.classList.contains('info-icon');
+  }
+  
+  function ensureInfoIconTooltip(icon) {
+    if (!isInfoIcon(icon)) {
+      return null;
+    }
+    let tooltip = icon.querySelector('.info-icon__tooltip');
+    if (!(tooltip instanceof HTMLElement)) {
+      tooltip = document.createElement('span');
+      tooltip.className = 'info-icon__tooltip';
+      tooltip.setAttribute('role', 'tooltip');
+      icon.append(tooltip);
+    }
+    let arrow = icon.querySelector('.info-icon__arrow');
+    if (!(arrow instanceof HTMLElement)) {
+      arrow = document.createElement('span');
+      arrow.className = 'info-icon__arrow';
+      arrow.setAttribute('aria-hidden', 'true');
+      icon.append(arrow);
+    }
+    const tooltipText = icon.dataset.tooltip || icon.getAttribute('aria-label') || '';
+    tooltip.textContent = tooltipText;
+    if (!tooltip.id) {
+      infoIconTooltipId += 1;
+      tooltip.id = `info-icon-tooltip-${infoIconTooltipId}`;
+    }
+    if (!icon.hasAttribute('aria-describedby')) {
+      icon.setAttribute('aria-describedby', tooltip.id);
+    }
+    return { tooltip, arrow };
+  }
+  
+  function positionInfoIconTooltip(icon) {
+    if (!isInfoIcon(icon)) {
+      return;
+    }
+    const parts = ensureInfoIconTooltip(icon);
+    if (!parts || !(parts.tooltip instanceof HTMLElement)) {
+      return;
+    }
+    icon.style.setProperty('--tooltip-shift', '0px');
+    const tooltipRect = parts.tooltip.getBoundingClientRect();
+    const viewportWidth = window.innerWidth || document.documentElement.clientWidth || 0;
+    const padding = TOOLTIP_VIEWPORT_PADDING;
+    let shift = 0;
+    if (tooltipRect.left < padding) {
+      shift = padding - tooltipRect.left;
+    } else if (tooltipRect.right > viewportWidth - padding) {
+      shift = -(tooltipRect.right - (viewportWidth - padding));
+    }
+    icon.style.setProperty('--tooltip-shift', `${shift}px`);
+  }
+  
+  function enhanceInfoIcon(icon) {
+    if (!isInfoIcon(icon)) {
+      return;
+    }
+    if (!icon.hasAttribute('tabindex')) {
+      icon.setAttribute('tabindex', '0');
+    }
+    icon.setAttribute('role', 'button');
+    icon.setAttribute('aria-expanded', icon.classList.contains(INFO_ICON_ACTIVE_CLASS) ? 'true' : 'false');
+    if (!icon.hasAttribute('aria-label') && icon.dataset.tooltip) {
+      icon.setAttribute('aria-label', icon.dataset.tooltip);
+    }
+    ensureInfoIconTooltip(icon);
+  }
+  
+  function enhanceAllInfoIcons() {
+    document.querySelectorAll('.info-icon').forEach(enhanceInfoIcon);
+  }
+  
+  function closeActiveInfoIcon() {
+    if (activeInfoIcon instanceof HTMLElement) {
+      activeInfoIcon.classList.remove(INFO_ICON_ACTIVE_CLASS);
+      activeInfoIcon.setAttribute('aria-expanded', 'false');
+    }
+    activeInfoIcon = null;
+  }
+  
+  function openInfoIcon(icon) {
+    if (!isInfoIcon(icon)) {
+      return;
+    }
+    if (activeInfoIcon && activeInfoIcon !== icon) {
+      closeActiveInfoIcon();
+    }
+    icon.classList.add(INFO_ICON_ACTIVE_CLASS);
+    icon.setAttribute('aria-expanded', 'true');
+    activeInfoIcon = icon;
+    positionInfoIconTooltip(icon);
+  }
+  
+  function toggleInfoIcon(icon) {
+    if (!isInfoIcon(icon)) {
+      return;
+    }
+    if (icon.classList.contains(INFO_ICON_ACTIVE_CLASS)) {
+      if (activeInfoIcon === icon) {
+        closeActiveInfoIcon();
+      } else {
+        icon.classList.remove(INFO_ICON_ACTIVE_CLASS);
+        icon.setAttribute('aria-expanded', 'false');
+      }
+    } else {
+      openInfoIcon(icon);
+    }
+  }
+  
+  document.addEventListener('click', event => {
+    const target = event.target instanceof HTMLElement ? event.target.closest('.info-icon') : null;
+    if (!isInfoIcon(target)) {
+      closeActiveInfoIcon();
+      return;
+    }
+    event.preventDefault();
+    enhanceInfoIcon(target);
+    toggleInfoIcon(target);
+    positionInfoIconTooltip(target);
+    if (typeof target.focus === 'function') {
+      target.focus();
+    }
+  });
+  
+  document.addEventListener('keydown', event => {
+    if (event.key === 'Escape') {
+      if (activeInfoIcon) {
+        event.preventDefault();
+        const current = activeInfoIcon;
+        closeActiveInfoIcon();
+        if (current instanceof HTMLElement && typeof current.focus === 'function') {
+          current.focus();
+        }
+      }
+      return;
+    }
+  
+    const target = event.target instanceof HTMLElement ? event.target.closest('.info-icon') : null;
+    if (!isInfoIcon(target)) {
+      return;
+    }
+  
+    if (event.key === 'Enter' || event.key === ' ') {
+      event.preventDefault();
+      enhanceInfoIcon(target);
+      toggleInfoIcon(target);
+      positionInfoIconTooltip(target);
+    }
+  });
+  
+  document.addEventListener('focusin', event => {
+    const target = event.target instanceof HTMLElement ? event.target.closest('.info-icon') : null;
+    if (!isInfoIcon(target)) {
+      closeActiveInfoIcon();
+      return;
+    }
+    enhanceInfoIcon(target);
+    positionInfoIconTooltip(target);
+  });
+  
+  enhanceAllInfoIcons();
+  
+  document.addEventListener('mouseover', event => {
+    const target = event.target instanceof HTMLElement ? event.target.closest('.info-icon') : null;
+    if (!isInfoIcon(target)) {
+      return;
+    }
+    positionInfoIconTooltip(target);
+  });
+  
+  window.addEventListener('resize', () => {
+    if (activeInfoIcon) {
+      positionInfoIconTooltip(activeInfoIcon);
+    }
+  });
+  
   function formatFixed(value, fractionDigits = 1) {
     const fixed = value.toFixed(fractionDigits);
     return fixed
       .replace(/\.0+$/, '')
       .replace(/(\.[0-9]*[1-9])0+$/, '$1');
   }
-
+  
   function parseNumber(value, fallback = 0, { min = -Infinity, max = Infinity } = {}) {
     if (value === null || value === undefined) {
       return fallback;
@@ -1461,23 +1962,23 @@ export function initCalculator() {
     }
     return Math.min(Math.max(parsed, min), max);
   }
-
+  
   function getCurrentTaxRate() {
     const percent = Math.min(Math.max(parseNumber(controls.taxRate?.value, 40), 0), 99.9);
     return percent / 100;
   }
-
+  
   function parseList(value) {
     if (!value) {
       return [];
     }
-
+  
     const values = [];
     const tokens = value
       .split(',')
       .map(token => token.trim())
       .filter(token => token.length > 0);
-
+  
     tokens.forEach(token => {
       const rangeMatch = token.match(/^(-?\d+(?:\.\d+)?)\s*-\s*(-?\d+(?:\.\d+)?)$/);
       if (rangeMatch) {
@@ -1486,7 +1987,7 @@ export function initCalculator() {
         if (!Number.isFinite(startValue) || !Number.isFinite(endValue)) {
           return;
         }
-
+  
         const rangeStart = Math.min(startValue, endValue);
         const rangeEnd = Math.max(startValue, endValue);
         for (let current = rangeStart; current <= rangeEnd; current += 1) {
@@ -1496,31 +1997,31 @@ export function initCalculator() {
         }
         return;
       }
-
+  
       const parsed = Math.round(Number(token));
       if (Number.isFinite(parsed) && parsed > 0) {
         values.push(parsed);
       }
     });
-
+  
     return Array.from(new Set(values)).sort((a, b) => a - b);
   }
-
+  
   function updateTablesLayout() {
     if (!tablesContainer) {
       return;
     }
-
+  
     const cards = tablesContainer.querySelectorAll('.card');
     if (cards.length <= 1) {
       tablesContainer.classList.remove('table-grid--stacked');
       return;
     }
-
+  
     const hasOverflow = tablesContainer.scrollWidth > tablesContainer.clientWidth + 1;
     tablesContainer.classList.toggle('table-grid--stacked', hasOverflow);
   }
-
+  
   function scheduleTablesLayoutUpdate() {
     if (!tablesContainer) {
       return;
@@ -1528,14 +2029,14 @@ export function initCalculator() {
     if (tablesLayoutUpdateScheduled) {
       return;
     }
-
+  
     tablesLayoutUpdateScheduled = true;
     requestAnimationFrame(() => {
       tablesLayoutUpdateScheduled = false;
       updateTablesLayout();
     });
   }
-
+  
   function getFixedCostTotal() {
     return Object.values(fixedCostFields).reduce((sum, fieldSet) => {
       if (!(fieldSet.annual instanceof HTMLInputElement)) {
@@ -1545,7 +2046,7 @@ export function initCalculator() {
       return sum + annualValue;
     }, 0);
   }
-
+  
   function updateFixedCostTotalDisplay() {
     const total = getFixedCostTotal();
     if (controls.fixedCosts instanceof HTMLInputElement) {
@@ -1553,7 +2054,7 @@ export function initCalculator() {
     }
     return total;
   }
-
+  
   function syncFixedCostPair(key, sourceType) {
     const fieldSet = fixedCostFields[key];
     if (!fieldSet) {
@@ -1564,23 +2065,23 @@ export function initCalculator() {
     if (!(source instanceof HTMLInputElement) || !(target instanceof HTMLInputElement)) {
       return;
     }
-
+  
     if (source.value === '') {
       target.value = '';
       updateFixedCostTotalDisplay();
       return;
     }
-
+  
     const parsed = Math.max(parseNumber(source.value, 0), 0);
     const derived = sourceType === 'monthly' ? parsed * 12 : parsed / 12;
-
+  
     if (Number.isFinite(derived)) {
       target.value = formatFixed(derived, 2);
     }
-
+  
     updateFixedCostTotalDisplay();
   }
-
+  
   Object.entries(fixedCostFields).forEach(([key, fieldSet]) => {
     const { monthly, annual } = fieldSet;
     if (monthly instanceof HTMLInputElement) {
@@ -1612,12 +2113,12 @@ export function initCalculator() {
       });
     }
   });
-
+  
   function getInputs() {
     const monthsOff = parseNumber(controls.monthsOff.value, 2, { min: 0, max: 12 });
     const weeksOffPerCycle = parseNumber(controls.weeksOffCycle.value, 1, { min: 0, max: 4 });
     const daysOffPerWeek = parseNumber(controls.daysOffWeek.value, 2, { min: 0, max: BASE_WORK_DAYS_PER_WEEK });
-
+  
     const taxRatePercent = Math.min(Math.max(parseNumber(controls.taxRate.value, 40), 0), 99.9);
     const taxRate = taxRatePercent / 100;
     if (desiredIncomeLockedAsGross) {
@@ -1644,7 +2145,7 @@ export function initCalculator() {
         }
       }
     }
-
+  
     let lessonPriceMin = null;
     if (controls.lessonPriceMin instanceof HTMLInputElement) {
       const rawMin = controls.lessonPriceMin.value;
@@ -1655,7 +2156,7 @@ export function initCalculator() {
         }
       }
     }
-
+  
     let lessonPriceMax = null;
     if (controls.lessonPriceMax instanceof HTMLInputElement) {
       const rawMax = controls.lessonPriceMax.value;
@@ -1666,13 +2167,13 @@ export function initCalculator() {
         }
       }
     }
-
+  
     if (lessonPriceMin != null && lessonPriceMax != null && lessonPriceMin > lessonPriceMax) {
       const tempMin = lessonPriceMax;
       lessonPriceMax = lessonPriceMin;
       lessonPriceMin = tempMin;
     }
-
+  
     const activeMonthShare = Math.min(Math.max((12 - monthsOff) / 12, 0), 1);
     const activeMonths = 12 * activeMonthShare;
     const weeksShare = Math.min(Math.max((4 - weeksOffPerCycle) / 4, 0), 1);
@@ -1682,24 +2183,24 @@ export function initCalculator() {
       Math.min(BASE_WORK_DAYS_PER_WEEK, BASE_WORK_DAYS_PER_WEEK - daysOffPerWeek)
     );
     const workingDaysPerYear = workingWeeks * workingDaysPerWeek;
-
+  
     lastActiveMonths = activeMonths;
     lastWorkingWeeks = workingWeeks;
     if (desiredIncomeLockedAsGross) {
       synchronizeLockedAcceptableIncomeNetValues(taxRate, activeMonths);
     }
-
+  
     const defaultTargetNetWeek = workingWeeks > 0 ? TARGET_NET_DEFAULT / workingWeeks : TARGET_NET_DEFAULT;
     const defaultTargetNetMonth = activeMonths > 0 ? TARGET_NET_DEFAULT / activeMonths : TARGET_NET_DEFAULT;
     const defaultTargetNetAverageWeek = TARGET_NET_DEFAULT / WEEKS_PER_YEAR;
     const defaultTargetNetAverageMonth = TARGET_NET_DEFAULT / MONTHS_PER_YEAR;
-
+  
     const storedYearNet = readDesiredIncomeNet('year', null);
     const storedWeekNet = readDesiredIncomeNet('week', null);
     const storedMonthNet = readDesiredIncomeNet('month', null);
     const storedAvgWeekNet = readDesiredIncomeNet('avgWeek', null);
     const storedAvgMonthNet = readDesiredIncomeNet('avgMonth', null);
-
+  
     const netYearValue = Number.isFinite(storedYearNet) ? Math.max(storedYearNet, 0) : TARGET_NET_DEFAULT;
     const netWeekValue = Number.isFinite(storedWeekNet)
       ? Math.max(storedWeekNet, 0)
@@ -1713,10 +2214,10 @@ export function initCalculator() {
     const netAvgMonthValue = Number.isFinite(storedAvgMonthNet)
       ? Math.max(storedAvgMonthNet, 0)
       : defaultTargetNetAverageMonth;
-
+  
     const hasWorkingWeeks = workingWeeks > 0;
     const hasActiveMonths = activeMonths > 0;
-
+  
     let targetNet;
     if (targetNetBasis === 'week') {
       targetNet = hasWorkingWeeks ? netWeekValue * workingWeeks : netYearValue;
@@ -1729,14 +2230,14 @@ export function initCalculator() {
     } else {
       targetNet = netYearValue;
     }
-
+  
     targetNet = Number.isFinite(targetNet) ? Math.max(targetNet, 0) : TARGET_NET_DEFAULT;
-
+  
     const targetNetPerWeek = hasWorkingWeeks ? targetNet / workingWeeks : null;
     const targetNetPerMonth = hasActiveMonths ? targetNet / activeMonths : null;
     const targetNetAveragePerWeek = targetNet / WEEKS_PER_YEAR;
     const targetNetAveragePerMonth = targetNet / MONTHS_PER_YEAR;
-
+  
     const derivedNetValues = {
       year: targetNet,
       week: targetNetPerWeek,
@@ -1744,7 +2245,7 @@ export function initCalculator() {
       avgWeek: targetNetAveragePerWeek,
       avgMonth: targetNetAveragePerMonth
     };
-
+  
     Object.entries(derivedNetValues).forEach(([key, value]) => {
       if (key === targetNetBasis) {
         if (!Number.isFinite(readDesiredIncomeNet(key, null)) && Number.isFinite(value)) {
@@ -1758,10 +2259,10 @@ export function initCalculator() {
         writeDesiredIncomeNet(key, null);
       }
     });
-
+  
     refreshDesiredIncomeDisplay(derivedNetValues, taxRate);
     refreshAcceptableIncomeDisplay(taxRate);
-
+  
     controls.taxRate.value = formatFixed(taxRate * 100, 1);
     controls.fixedCosts.value = formatFixed(fixedCosts, 2);
     if (controls.variableCostPerClass instanceof HTMLInputElement) {
@@ -1801,10 +2302,10 @@ export function initCalculator() {
         controls.lessonCost.value = '';
       }
     }
-
+  
     controls.workingWeeksDisplay.textContent = formatFixed(workingWeeks, 2);
     controls.workingDaysDisplay.textContent = formatFixed(workingDaysPerYear, 2);
-
+  
     return {
       targetNet,
       targetNetPerWeek,
@@ -1837,7 +2338,7 @@ export function initCalculator() {
       weeksShare
     };
   }
-
+  
   function formatCurrency(symbol, value) {
     if (!Number.isFinite(value)) {
       return `${symbol}0`;
@@ -1846,7 +2347,7 @@ export function initCalculator() {
     const formatted = numberFormatter.format(Math.abs(rounded));
     return rounded < 0 ? `-${symbol}${formatted}` : `${symbol}${formatted}`;
   }
-
+  
   function formatCurrencyDetailed(symbol, value, digits = 2) {
     if (!Number.isFinite(value)) {
       return `${symbol}0.00`;
@@ -1858,21 +2359,21 @@ export function initCalculator() {
     });
     return value < 0 ? `-${symbol}${formatted}` : `${symbol}${formatted}`;
   }
-
+  
   function formatCurrencyOrDash(symbol, value, digits = 2) {
-    return Number.isFinite(value) ? formatCurrencyDetailed(symbol, value, digits) : '—';
+    return Number.isFinite(value) ? formatCurrencyDetailed(symbol, value, digits) : '–';
   }
-
+  
   function formatNumberValue(value, maximumFractionDigits = 2) {
     if (!Number.isFinite(value)) {
-      return '—';
+      return '–';
     }
     return value.toLocaleString(undefined, {
       minimumFractionDigits: 0,
       maximumFractionDigits
     });
   }
-
+  
   function escapeHtml(value) {
     if (value === null || value === undefined) {
       return '';
@@ -1884,7 +2385,7 @@ export function initCalculator() {
       .replace(/"/g, '&quot;')
       .replace(/'/g, '&#39;');
   }
-
+  
   function cloneInputs(inputs) {
     if (!inputs || typeof inputs !== 'object') {
       return null;
@@ -1917,36 +2418,36 @@ export function initCalculator() {
       targetNetAveragePerMonth: inputs.targetNetAveragePerMonth
     };
   }
-
+  
   function findBreakdownButton(context) {
     if (!context || !tablesContainer) {
       return null;
     }
-
+  
     const { rowIndex, columnIndex, variant } = context;
     if (!Number.isFinite(rowIndex) || !Number.isFinite(columnIndex)) {
       return null;
     }
     const variantValue = variant === 'base' ? 'base' : 'buffered';
-
+  
     return tablesContainer.querySelector(
       `button.price-line[data-row="${rowIndex}"][data-column="${columnIndex}"][data-variant="${variantValue}"]`
     );
   }
-
+  
   function formatBreakdownCurrency(value) {
-    return Number.isFinite(value) ? formatCurrency(latestCurrencySymbol, value) : '—';
+    return Number.isFinite(value) ? formatCurrency(latestCurrencySymbol, value) : '–';
   }
-
+  
   function findBestPricingCombination(studentsTarget, classesPerWeekTarget, pricingData = latestPricingData) {
     if (!Array.isArray(pricingData) || !pricingData.length) {
       return null;
     }
-
+  
     const tolerance = 1e-6;
     let bestMatch = null;
     let bestScore = Infinity;
-
+  
     pricingData.forEach(row => {
       if (!row || !Array.isArray(row.columns)) {
         return;
@@ -1974,11 +2475,11 @@ export function initCalculator() {
         }
       });
     });
-
+  
     if (!bestMatch) {
       return null;
     }
-
+  
     return {
       row: bestMatch.row,
       column: bestMatch.column,
@@ -1986,7 +2487,7 @@ export function initCalculator() {
       exactClasses: bestMatch.classDiff < tolerance
     };
   }
-
+  
   function buildAccountingReport({
     studentsRequested,
     classesPerWeekRequested,
@@ -2001,14 +2502,14 @@ export function initCalculator() {
     if (!combination || !combination.row || !combination.column) {
       return null;
     }
-
+  
     const { row, column, exactStudents, exactClasses } = combination;
     const breakdown = column.base && column.base.breakdown ? column.base.breakdown : null;
-
+  
     if (!breakdown || !breakdown.perLesson || !breakdown.totals) {
       return null;
     }
-
+  
     const totals = breakdown.totals;
     const studentsUsed = Number.isFinite(row.students) ? row.students : studentsRequested;
     const classesPerWeekUsed = Number.isFinite(column.classesPerWeek)
@@ -2025,7 +2526,7 @@ export function initCalculator() {
     const hoursPerClass = Number.isFinite(hoursPerClassRequested) ? hoursPerClassRequested : inputs?.hoursPerLesson;
     const weeklyHours = Number.isFinite(hoursPerClass) ? hoursPerClass * classesPerWeekUsed : null;
     const monthlyHours = Number.isFinite(hoursPerClass) ? hoursPerClass * classesPerMonth : null;
-
+  
     const revenuePerLessonInclVat = Number.isFinite(totals.priceInclVatPerLesson)
       ? totals.priceInclVatPerLesson
       : totals.priceInclVatPerStudent * studentsUsed;
@@ -2035,7 +2536,7 @@ export function initCalculator() {
     const monthlyRevenueInclVat = revenuePerLessonInclVat * classesPerMonth;
     const monthlyRevenueExVat = revenuePerLessonExVat * classesPerMonth;
     const monthlyVat = monthlyRevenueInclVat - monthlyRevenueExVat;
-
+  
     const variableCostPerClass = Number.isFinite(inputs?.variableCostPerClass)
       ? inputs.variableCostPerClass
       : 0;
@@ -2052,24 +2553,24 @@ export function initCalculator() {
     const perLessonVariableClass = variableCostPerClass;
     const perLessonVariableStudent = variableCostPerStudent * studentsUsed;
     const perLessonVariableMonthly = classesPerMonth > 0 ? monthlyVariableStudentMonthly / classesPerMonth : 0;
-
+  
     const fixedCosts = Number.isFinite(inputs?.fixedCosts) ? inputs.fixedCosts : 0;
     const monthlyFixedCosts = activeMonths > 0 ? fixedCosts / activeMonths : 0;
     const perLessonFixed = classesPerMonth > 0 ? monthlyFixedCosts / classesPerMonth : 0;
-
+  
     const taxRate = Number.isFinite(inputs?.taxRate) ? inputs.taxRate : 0;
     const profitBeforeTax = monthlyRevenueExVat - monthlyVariableTotal - monthlyFixedCosts;
     const monthlyIncomeTax = profitBeforeTax > 0 ? profitBeforeTax * taxRate : 0;
     const perLessonIncomeTax = classesPerMonth > 0 ? monthlyIncomeTax / classesPerMonth : 0;
     const monthlyNetIncome = profitBeforeTax - monthlyIncomeTax;
     const perLessonNetIncome = classesPerMonth > 0 ? monthlyNetIncome / classesPerMonth : 0;
-
+  
     const perLessonVat = classesPerMonth > 0 ? monthlyVat / classesPerMonth : 0;
     const perLessonOutgoings =
       perLessonVat + perLessonVariableClass + perLessonVariableStudent + perLessonVariableMonthly + perLessonFixed + perLessonIncomeTax;
     const monthlyOutgoings =
       monthlyVat + monthlyVariableClass + monthlyVariableStudent + monthlyVariableStudentMonthly + monthlyFixedCosts + monthlyIncomeTax;
-
+  
     const netMargin = monthlyRevenueExVat > 0 ? (monthlyNetIncome / monthlyRevenueExVat) * 100 : null;
     const effectiveHourlyNet = Number.isFinite(monthlyHours) && monthlyHours > 0
       ? monthlyNetIncome / monthlyHours
@@ -2080,7 +2581,7 @@ export function initCalculator() {
     const approxWeeksPerActiveMonth = activeMonths > 0 && Number.isFinite(workingWeeks)
       ? workingWeeks / activeMonths
       : 0;
-
+  
     const currency = typeof currencySymbol === 'string' && currencySymbol.trim() !== ''
       ? currencySymbol
       : '€';
@@ -2089,7 +2590,7 @@ export function initCalculator() {
       dateStyle: 'long',
       timeStyle: 'short'
     });
-
+  
     const notes = [];
     if (!exactStudents && Number.isFinite(studentsRequested)) {
       notes.push(
@@ -2101,7 +2602,7 @@ export function initCalculator() {
         `Requested ${formatNumberValue(classesPerWeekRequested, 2)} classes per week; closest available combination uses ${formatNumberValue(classesPerWeekUsed, 2)}.`
       );
     }
-
+  
     if (pricingMode === PRICING_MODE_TARGET) {
       if (Number.isFinite(bufferPercent) && bufferPercent > 0 && Number.isFinite(bufferedMonthlyNet)) {
         notes.push(
@@ -2113,14 +2614,14 @@ export function initCalculator() {
     } else {
       notes.push('Report uses the manual lesson price you entered.');
     }
-
+  
     const notesMarkup = notes.length
       ? `<section class="report-section">
           <h2>Notes</h2>
           <ul class="report-list">${notes.map(note => `<li>${escapeHtml(note)}</li>`).join('')}</ul>
         </section>`
       : '';
-
+  
     const summaryItems = [
       { label: 'Students per class', value: formatNumberValue(studentsUsed, 2) },
       { label: 'Classes per week', value: formatNumberValue(classesPerWeekUsed, 2) },
@@ -2133,7 +2634,7 @@ export function initCalculator() {
       { label: 'VAT rate', value: `${formatFixed((inputs?.vatRate ?? 0) * 100, 1)}%` },
       { label: 'Income tax rate', value: `${formatFixed(taxRate * 100, 1)}%` }
     ];
-
+  
     const summaryMarkup = summaryItems
       .map(item => `
         <div class="summary-item">
@@ -2142,9 +2643,9 @@ export function initCalculator() {
         </div>
       `)
       .join('');
-
+  
     const currencyEscaped = escapeHtml(currency);
-
+  
     const html = `<!DOCTYPE html>
   <html lang="en">
     <head>
@@ -2353,17 +2854,17 @@ export function initCalculator() {
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, monthlyVat))}</td>
           </tr>
           <tr>
-            <th scope="row">Variable costs — per class</th>
+            <th scope="row">Variable costs – per class</th>
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, perLessonVariableClass))}</td>
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, monthlyVariableClass))}</td>
           </tr>
           <tr>
-            <th scope="row">Variable costs — per student per class</th>
+            <th scope="row">Variable costs – per student per class</th>
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, perLessonVariableStudent))}</td>
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, monthlyVariableStudent))}</td>
           </tr>
           <tr>
-            <th scope="row">Variable costs — per student per month</th>
+            <th scope="row">Variable costs – per student per month</th>
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, perLessonVariableMonthly))}</td>
             <td class="numeric">${escapeHtml(formatCurrencyOrDash(currency, monthlyVariableStudentMonthly))}</td>
           </tr>
@@ -2405,15 +2906,15 @@ export function initCalculator() {
           </tr>
           <tr>
             <th scope="row">Net margin (on ex-VAT revenue)</th>
-            <td class="numeric">${Number.isFinite(netMargin) ? escapeHtml(`${formatFixed(netMargin, 1)}%`) : '—'}</td>
+            <td class="numeric">${Number.isFinite(netMargin) ? escapeHtml(`${formatFixed(netMargin, 1)}%`) : '–'}</td>
           </tr>
           <tr>
             <th scope="row">Gross hourly revenue</th>
-            <td class="numeric">${Number.isFinite(effectiveHourlyGross) ? escapeHtml(formatCurrencyDetailed(currency, effectiveHourlyGross)) : '—'}</td>
+            <td class="numeric">${Number.isFinite(effectiveHourlyGross) ? escapeHtml(formatCurrencyDetailed(currency, effectiveHourlyGross)) : '–'}</td>
           </tr>
           <tr>
             <th scope="row">Net hourly pay</th>
-            <td class="numeric">${Number.isFinite(effectiveHourlyNet) ? escapeHtml(formatCurrencyDetailed(currency, effectiveHourlyNet)) : '—'}</td>
+            <td class="numeric">${Number.isFinite(effectiveHourlyNet) ? escapeHtml(formatCurrencyDetailed(currency, effectiveHourlyNet)) : '–'}</td>
           </tr>
           <tr>
             <th scope="row">Net income per lesson</th>
@@ -2427,10 +2928,10 @@ export function initCalculator() {
   </main>
     </body>
   </html>`;
-
+  
     return html;
   }
-
+  
   function populateBreakdownDialog(context) {
     if (
       !context ||
@@ -2440,49 +2941,49 @@ export function initCalculator() {
     ) {
       return false;
     }
-
+  
     const isTargetMode = latestPricingMode === PRICING_MODE_TARGET;
     const isLessonMode = latestPricingMode === PRICING_MODE_LESSON;
-
+  
     if (!isTargetMode && !isLessonMode) {
       return false;
     }
-
+  
     const { rowIndex, columnIndex, variant } = context;
     const row = latestPricingData[rowIndex];
     if (!row || !Array.isArray(row.columns)) {
       return false;
     }
-
+  
     const column = row.columns[columnIndex];
     if (!column) {
       return false;
     }
-
+  
     const pricingVariant = variant === 'base' ? column.base : column.buffered;
     if (!pricingVariant || !pricingVariant.breakdown) {
       return false;
     }
-
+  
     const breakdown = pricingVariant.breakdown;
     const perStudentInclVat = formatCurrency(latestCurrencySymbol, pricingVariant.priceInclVat);
     const perStudentExVat = formatCurrency(latestCurrencySymbol, pricingVariant.priceExVat);
     const perLessonInclVat = formatCurrency(latestCurrencySymbol, breakdown.totals.priceInclVatPerLesson);
     const perLessonExVat = formatCurrency(latestCurrencySymbol, breakdown.totals.priceExVatPerLesson);
-
+  
     const studentsDisplay = Number.isFinite(row.students)
       ? numberFormatter.format(row.students)
-      : '—';
+      : '–';
     const classesPerWeekValue = column.classesPerWeek;
     const classesPerWeekDisplay = Number.isFinite(classesPerWeekValue)
       ? (Number.isInteger(classesPerWeekValue)
           ? numberFormatter.format(classesPerWeekValue)
           : formatFixed(classesPerWeekValue, 2))
-      : '—';
+      : '–';
     const classesPerYearDisplay = Number.isFinite(column.classesPerYear)
       ? numberFormatter.format(Math.round(column.classesPerYear))
-      : '—';
-
+      : '–';
+  
     let variantLabel;
     if (isLessonMode) {
       variantLabel = variant === 'base'
@@ -2493,23 +2994,23 @@ export function initCalculator() {
         ? 'Base price (no buffer)'
         : `Buffered price (+${formatFixed(latestBufferPercent, 1)}% extra safety margin)`;
     }
-
+  
     if (breakdownVariant) {
       breakdownVariant.textContent = variantLabel;
     }
-
+  
     if (breakdownPrice) {
       breakdownPrice.textContent = `${perStudentInclVat} per student (${perStudentExVat} ex VAT)`;
     }
-
+  
     if (breakdownMeta) {
       breakdownMeta.textContent = `${studentsDisplay} students · ${classesPerWeekDisplay} classes/week (≈ ${classesPerYearDisplay} / yr)`;
     }
-
+  
     if (breakdownTotal) {
       breakdownTotal.textContent = `Total collected per lesson: ${perLessonInclVat} (${perLessonExVat} ex VAT)`;
     }
-
+  
     const tableRows = [
       { key: 'vat', label: 'VAT remitted' },
       { key: 'variableCosts', label: 'Variable costs' },
@@ -2517,7 +3018,7 @@ export function initCalculator() {
       { key: 'incomeTax', label: 'Income tax' },
       { key: 'netIncome', label: 'Net income after tax' }
     ];
-
+  
     if (breakdownTableBody) {
       const rowsMarkup = tableRows
         .map(item => {
@@ -2534,7 +3035,7 @@ export function initCalculator() {
         .join('');
       breakdownTableBody.innerHTML = rowsMarkup;
     }
-
+  
     const legendItems = [
       { key: 'vat', label: 'VAT remitted', value: breakdown.perLesson.vat, color: 'var(--breakdown-color-vat)' },
       { key: 'variableCosts', label: 'Variable costs', value: breakdown.perLesson.variableCosts, color: 'var(--breakdown-color-variable)' },
@@ -2542,7 +3043,7 @@ export function initCalculator() {
       { key: 'incomeTax', label: 'Income tax', value: breakdown.perLesson.incomeTax, color: 'var(--breakdown-color-tax)' },
       { key: 'netIncome', label: 'Net income after tax', value: breakdown.perLesson.netIncome, color: 'var(--breakdown-color-net)' }
     ];
-
+  
     if (breakdownLegend) {
       const legendMarkup = legendItems
         .map(item => `
@@ -2555,11 +3056,11 @@ export function initCalculator() {
         .join('');
       breakdownLegend.innerHTML = legendMarkup;
     }
-
+  
     if (breakdownChartGraphic) {
       const positiveSegments = legendItems.map(item => Math.max(item.value, 0));
       const total = positiveSegments.reduce((sum, value) => sum + value, 0);
-
+  
       if (total > 0) {
         let currentAngle = 0;
         const segments = legendItems
@@ -2574,40 +3075,40 @@ export function initCalculator() {
             return `${item.color} ${start}deg ${end}deg`;
           })
           .filter(Boolean);
-
+  
         breakdownChartGraphic.style.background = segments.length
           ? `conic-gradient(${segments.join(', ')})`
           : 'radial-gradient(circle at 50% 50%, var(--price-line-bg), var(--row-header-bg))';
       } else {
         breakdownChartGraphic.style.background = 'radial-gradient(circle at 50% 50%, var(--price-line-bg), var(--row-header-bg))';
       }
-
+  
     }
-
+  
     return true;
   }
-
+  
   function openBreakdownDialog(context, triggerElement) {
     if (!breakdownDialog) {
       return;
     }
-
+  
     const normalizedContext = {
       rowIndex: Number(context?.rowIndex),
       columnIndex: Number(context?.columnIndex),
       variant: context?.variant === 'base' ? 'base' : 'buffered'
     };
-
+  
     if (!populateBreakdownDialog(normalizedContext)) {
       return;
     }
-
+  
     activeBreakdownContext = normalizedContext;
     breakdownTriggerElement = triggerElement instanceof HTMLElement ? triggerElement : null;
     breakdownDialog.hidden = false;
     breakdownDialog.scrollTop = 0;
     setBodyScrollLock('breakdown', true);
-
+  
     window.requestAnimationFrame(() => {
       const focusable = getFocusableElements(breakdownDialog);
       if (focusable.length) {
@@ -2616,49 +3117,49 @@ export function initCalculator() {
         breakdownClose.focus();
       }
     });
-
+  
     document.addEventListener('keydown', handleBreakdownKeydown);
   }
-
+  
   function closeBreakdownDialog() {
     if (!breakdownDialog || breakdownDialog.hidden) {
       return;
     }
-
+  
     breakdownDialog.hidden = true;
     setBodyScrollLock('breakdown', false);
     document.removeEventListener('keydown', handleBreakdownKeydown);
-
+  
     const trigger = breakdownTriggerElement;
     breakdownTriggerElement = null;
     activeBreakdownContext = null;
-
+  
     if (trigger instanceof HTMLElement && trigger.isConnected && typeof trigger.focus === 'function') {
       trigger.focus();
     }
   }
-
+  
   function handleBreakdownKeydown(event) {
     if (!breakdownDialog || breakdownDialog.hidden) {
       return;
     }
-
+  
     if (event.key === 'Escape') {
       event.preventDefault();
       closeBreakdownDialog();
       return;
     }
-
+  
     if (event.key === 'Tab') {
       const focusable = getFocusableElements(breakdownDialog);
       if (!focusable.length) {
         event.preventDefault();
         return;
       }
-
+  
       const first = focusable[0];
       const last = focusable[focusable.length - 1];
-
+  
       if (!event.shiftKey && document.activeElement === last) {
         event.preventDefault();
         first.focus();
@@ -2668,12 +3169,12 @@ export function initCalculator() {
       }
     }
   }
-
+  
   function refreshBreakdownDialog() {
     if (!breakdownDialog || breakdownDialog.hidden || !activeBreakdownContext) {
       return false;
     }
-
+  
     const refreshed = populateBreakdownDialog(activeBreakdownContext);
     if (refreshed) {
       const updatedTrigger = findBreakdownButton(activeBreakdownContext);
@@ -2681,10 +3182,10 @@ export function initCalculator() {
         breakdownTriggerElement = updatedTrigger;
       }
     }
-
+  
     return refreshed;
   }
-
+  
   function computeNetIncomeFromRevenue(revenue, fixedCosts, effectiveTaxRate, variableCosts = 0) {
     if (!Number.isFinite(revenue)) {
       return null;
@@ -2693,7 +3194,7 @@ export function initCalculator() {
     const profitBeforeTax = revenue - fixedCosts - normalizedVariableCosts;
     return profitBeforeTax * (1 - effectiveTaxRate);
   }
-
+  
   function shouldHighlightIncome({ monthlyNet, annualNet }, options = {}) {
     const {
       acceptableIncome = null,
@@ -2701,30 +3202,30 @@ export function initCalculator() {
       taxRate = 0,
       activeMonths = MONTHS_PER_YEAR
     } = options;
-
+  
     if (!acceptableIncome || typeof acceptableIncome !== 'object') {
       return false;
     }
-
+  
     const { basis, minAnnualNet, maxAnnualNet } = acceptableIncome;
     const hasMin = Number.isFinite(minAnnualNet);
     const hasMax = Number.isFinite(maxAnnualNet);
-
+  
     if (!basis || (!hasMin && !hasMax)) {
       return false;
     }
-
+  
     const normalizedTaxRate = Math.min(Math.max(taxRate, 0), 0.9999);
     const denominator = Math.max(1 - normalizedTaxRate, 0.0001);
     const monthsForRange = activeMonths > 0 ? activeMonths : MONTHS_PER_YEAR;
-
+  
     const convertNetToDisplay = value => {
       if (!Number.isFinite(value)) {
         return null;
       }
       return displayMode === 'gross' ? value / denominator : value;
     };
-
+  
     if (basis === 'annual') {
       if (!Number.isFinite(annualNet)) {
         return false;
@@ -2743,7 +3244,7 @@ export function initCalculator() {
       }
       return true;
     }
-
+  
     if (!Number.isFinite(monthlyNet)) {
       return false;
     }
@@ -2761,12 +3262,12 @@ export function initCalculator() {
     }
     return true;
   }
-
+  
   function buildPricingTable(data, symbol, bufferPercent, options = {}) {
     if (!data.length) {
       return `<div class="card"><p class="status-message">No valid combinations available.</p></div>`;
     }
-
+  
     const {
       mode = PRICING_MODE_TARGET,
       showBasePrices: showBase = false,
@@ -2784,10 +3285,10 @@ export function initCalculator() {
     if (showBasePricesActive) {
       cardClasses.push('show-base-prices');
     }
-
+  
     const hasPreferredRange =
       Number.isFinite(minLessonPrice) || Number.isFinite(maxLessonPrice);
-
+  
     const convertNetIncomeForDisplay = value => {
       if (!Number.isFinite(value)) {
         return null;
@@ -2796,15 +3297,15 @@ export function initCalculator() {
         ? convertNetToGross(value, incomeTaxRate)
         : value;
     };
-
+  
     const formatIncomeForDisplay = value => {
       const converted = convertNetIncomeForDisplay(value);
-      return Number.isFinite(converted) ? formatCurrency(symbol, converted) : '—';
+      return Number.isFinite(converted) ? formatCurrency(symbol, converted) : '–';
     };
-
+  
     const monthlyIncomeLabel = incomeDisplayMode === 'gross' ? 'Monthly gross' : 'Monthly net';
     const annualIncomeLabel = incomeDisplayMode === 'gross' ? 'Annual gross' : 'Annual net';
-
+  
     const isPriceOutOfRange = value => {
       if (!hasPreferredRange || !Number.isFinite(value)) {
         return false;
@@ -2817,10 +3318,10 @@ export function initCalculator() {
       }
       return false;
     };
-
+  
     const normalizedLessonHours = Number.isFinite(lessonHours) && lessonHours > 0 ? lessonHours : null;
     const hourlyRateLabel = incomeDisplayMode === 'gross' ? 'Hourly gross' : 'Hourly net';
-
+  
     const computeHourlyRate = breakdown => {
       if (!normalizedLessonHours) {
         return null;
@@ -2838,12 +3339,12 @@ export function initCalculator() {
       const perHour = lessonValue / normalizedLessonHours;
       return Number.isFinite(perHour) ? perHour : null;
     };
-
+  
     const formatHourlyRateDisplay = breakdown => {
       const rate = computeHourlyRate(breakdown);
-      return Number.isFinite(rate) ? formatCurrency(symbol, rate) : '—';
+      return Number.isFinite(rate) ? formatCurrency(symbol, rate) : '–';
     };
-
+  
     const toggleMarkup = mode === PRICING_MODE_TARGET
       ? `<div class="price-toggle-row">
           <button type="button" class="price-reveal-toggle secondary" aria-expanded="${showBasePricesActive}">
@@ -2851,13 +3352,13 @@ export function initCalculator() {
           </button>
         </div>`
       : '';
-
+  
     const columnHeaders = data[0].columns
       .map(col => {
         return `<th scope="col">${col.classesPerWeek}/week<span class="sub-label">≈ ${numberFormatter.format(col.classesPerYear)} / yr</span></th>`;
       })
       .join('');
-
+  
     const rowsHtml = data
       .map((row, rowIndex) => {
         const cells = row.columns
@@ -2869,7 +3370,7 @@ export function initCalculator() {
               const bufferedAnnualDisplay = formatIncomeForDisplay(col.manualNet.bufferedAnnual);
               const bufferedHourlyDisplay = formatHourlyRateDisplay(col.buffered.breakdown);
               const baseHourlyDisplay = formatHourlyRateDisplay(col.base.breakdown);
-
+  
               const highlightBase = shouldHighlightIncome(
                 { monthlyNet: col.manualNet.monthly, annualNet: col.manualNet.annual },
                 {
@@ -2891,12 +3392,12 @@ export function initCalculator() {
                   activeMonths: incomeActiveMonths
                 }
               );
-
+  
               const baseButtonClass = highlightBase ? 'price-line price-line--acceptable' : 'price-line';
               const bufferedButtonClass = highlightBuffered
                 ? 'price-line buffered price-line--acceptable'
                 : 'price-line buffered';
-
+  
               return `
               <td>
                 <div class="price-pair">
@@ -2928,7 +3429,7 @@ export function initCalculator() {
               </td>
             `;
             }
-
+  
             const bufferedExVat = formatCurrency(symbol, col.buffered.priceExVat);
             const bufferedInclVat = formatCurrency(symbol, col.buffered.priceInclVat);
             const baseExVat = formatCurrency(symbol, col.base.priceExVat);
@@ -3015,7 +3516,7 @@ export function initCalculator() {
         return `<tr><th scope="row">${row.students}<span class="sub-label">students</span></th>${cells}</tr>`;
       })
       .join('');
-
+  
     return `
       <div class="${cardClasses.join(' ')}">
         ${toggleMarkup}
@@ -3036,7 +3537,7 @@ export function initCalculator() {
       </div>
     `;
   }
-
+  
   function buildActiveTargetsTable({
     currencySymbol,
     revenueNeeded,
@@ -3047,7 +3548,7 @@ export function initCalculator() {
     if (!Number.isFinite(revenueNeeded) || revenueNeeded <= 0) {
       return '';
     }
-
+  
     const targets = [
       {
         label: 'Active day',
@@ -3065,18 +3566,18 @@ export function initCalculator() {
         suffix: 'active months / year'
       }
     ];
-
+  
     const rows = targets
       .map(target => {
         const hasBasis = Number.isFinite(target.basis) && target.basis > 0;
         const amount = hasBasis ? revenueNeeded / target.basis : null;
         const amountDisplay = hasBasis && Number.isFinite(amount)
           ? formatCurrency(currencySymbol, amount)
-          : '—';
+          : '–';
         const basisDisplay = hasBasis
           ? `Based on ≈ ${formatFixed(target.basis, 2)} ${target.suffix}`
-          : '—';
-
+          : '–';
+  
         return `
           <tr>
             <th scope="row">${target.label}</th>
@@ -3086,7 +3587,7 @@ export function initCalculator() {
         `;
       })
       .join('');
-
+  
     return `
       <div class="card">
         <div class="card-scroll">
@@ -3117,7 +3618,7 @@ export function initCalculator() {
       </div>
     `;
   }
-
+  
   function computeTables(inputs) {
     const {
       targetNet,
@@ -3136,12 +3637,12 @@ export function initCalculator() {
       lessonCostInclVat,
       activeMonths
     } = inputs;
-
+  
     latestResults = [];
-
+  
     const manualMode = Number.isFinite(lessonCostInclVat);
     const mode = manualMode ? PRICING_MODE_LESSON : PRICING_MODE_TARGET;
-
+  
     const pricingData = [];
     const effectiveTaxRate = Math.min(taxRate, 0.99);
     const hasSchedule =
@@ -3149,28 +3650,28 @@ export function initCalculator() {
       studentsPerClass.length &&
       Number.isFinite(workingWeeks) &&
       workingWeeks > 0;
-
+  
     const hasWorkingWeeks = Number.isFinite(workingWeeks) && workingWeeks > 0;
     const hasActiveMonths = Number.isFinite(activeMonths) && activeMonths > 0;
-
+  
     let manualNetSummary = null;
     let revenueNeeded = null;
-
+  
     if (!hasSchedule) {
       latestResultsMode = mode;
       return { pricingData, bufferPercent, revenueNeeded, mode, manualNetSummary };
     }
-
+  
     const sortedStudents = [...studentsPerClass];
     sortedStudents.sort((a, b) => a - b);
-
+  
     const normalizedActiveMonths = Number.isFinite(activeMonths) && activeMonths > 0 ? activeMonths : 0;
-
+  
     const columnsMeta = classesPerWeek.map(classes => {
       const classesPerYear = classes * workingWeeks;
       return { classesPerWeek: classes, classesPerYear };
     });
-
+  
     function buildPriceBreakdown({ priceExVatValue, priceInclVatValue, studentCount, classesPerYearValue }) {
       const normalizedPriceExVat = Number.isFinite(priceExVatValue) ? priceExVatValue : 0;
       const normalizedPriceInclVat = Number.isFinite(priceInclVatValue) ? priceInclVatValue : 0;
@@ -3188,7 +3689,7 @@ export function initCalculator() {
       const profitBeforeTaxPerLesson = revenueExVatPerLesson - variableCostsPerLesson - fixedAllocationPerLesson;
       const incomeTaxPerLesson = profitBeforeTaxPerLesson > 0 ? profitBeforeTaxPerLesson * effectiveTaxRate : 0;
       const netIncomePerLesson = profitBeforeTaxPerLesson - incomeTaxPerLesson;
-
+  
       return {
         perLesson: {
           vat: vatPerLesson,
@@ -3214,17 +3715,17 @@ export function initCalculator() {
         }
       };
     }
-
+  
     if (mode === PRICING_MODE_LESSON) {
       const vatDivisor = Math.max(1 + vatRate, 0.0001);
       const priceInclVatPerStudent = lessonCostInclVat;
       const priceExVatPerStudent = lessonCostInclVat / vatDivisor;
       const attendanceMultiplier = Math.max(1 - buffer, 0);
       const formattedBuffer = formatFixed(bufferPercent, 1);
-
+  
       for (const students of sortedStudents) {
         const row = { students, columns: [] };
-
+  
         for (const column of columnsMeta) {
           const classesPerYear = column.classesPerYear;
           const annualRevenue = priceExVatPerStudent * students * classesPerYear;
@@ -3233,7 +3734,7 @@ export function initCalculator() {
             variableCostPerClass * classesPerYear +
             variableCostPerStudent * students * classesPerYear +
             variableCostPerStudentMonthly * students * normalizedActiveMonths;
-
+  
           const annualNet = computeNetIncomeFromRevenue(
             annualRevenue,
             fixedCosts,
@@ -3246,14 +3747,14 @@ export function initCalculator() {
             effectiveTaxRate,
             annualVariableCosts
           );
-
+  
           const monthlyNet = Number.isFinite(annualNet) && hasActiveMonths
             ? annualNet / activeMonths
             : null;
           const bufferedMonthlyNet = Number.isFinite(bufferedAnnualNet) && hasActiveMonths
             ? bufferedAnnualNet / activeMonths
             : null;
-
+  
           if (!manualNetSummary && Number.isFinite(annualNet)) {
             manualNetSummary = {
               annual: annualNet,
@@ -3263,14 +3764,14 @@ export function initCalculator() {
               averageMonthly: annualNet / MONTHS_PER_YEAR
             };
           }
-
+  
           const baseBreakdown = buildPriceBreakdown({
             priceExVatValue: priceExVatPerStudent,
             priceInclVatValue: priceInclVatPerStudent,
             studentCount: students,
             classesPerYearValue: classesPerYear
           });
-
+  
           row.columns.push({
             classesPerWeek: column.classesPerWeek,
             classesPerYear,
@@ -3291,7 +3792,7 @@ export function initCalculator() {
               breakdown: baseBreakdown
             }
           });
-
+  
           latestResults.push({
             variant: 'Full attendance',
             students,
@@ -3309,26 +3810,26 @@ export function initCalculator() {
             annualNet: bufferedAnnualNet
           });
         }
-
+  
         pricingData.push(row);
       }
-
+  
       latestResultsMode = mode;
       return { pricingData, bufferPercent, revenueNeeded, mode, manualNetSummary };
     }
-
+  
     const denominator = Math.max(1 - effectiveTaxRate, 0.0001);
     const profitBeforeTax = targetNet / denominator;
     revenueNeeded = profitBeforeTax + fixedCosts;
-
+  
     if (!Number.isFinite(revenueNeeded) || revenueNeeded <= 0) {
       latestResultsMode = mode;
       return { pricingData, bufferPercent, revenueNeeded, mode, manualNetSummary };
     }
-
+  
     for (const students of sortedStudents) {
       const row = { students, columns: [] };
-
+  
       for (const column of columnsMeta) {
         const classesPerYear = column.classesPerYear;
         const effectiveClassesPerYear = Math.max(classesPerYear, 1);
@@ -3342,12 +3843,12 @@ export function initCalculator() {
         const bufferedExVat = priceExVat * (1 + buffer);
         const priceInclVat = priceExVat * (1 + vatRate);
         const bufferedInclVat = bufferedExVat * (1 + vatRate);
-
+  
         const roundedBaseExVat = Math.round(priceExVat);
         const roundedBaseInclVat = Math.round(priceInclVat);
         const roundedBufferedExVat = Math.round(bufferedExVat);
         const roundedBufferedInclVat = Math.round(bufferedInclVat);
-
+  
         const baseBreakdown = buildPriceBreakdown({
           priceExVatValue: priceExVat,
           priceInclVatValue: priceInclVat,
@@ -3360,7 +3861,7 @@ export function initCalculator() {
           studentCount: students,
           classesPerYearValue: classesPerYear
         });
-
+  
         const baseAnnualNet = Number.isFinite(baseBreakdown?.perLesson?.netIncome)
           ? baseBreakdown.perLesson.netIncome * classesPerYear
           : null;
@@ -3373,7 +3874,7 @@ export function initCalculator() {
         const bufferedMonthlyNet = Number.isFinite(bufferedAnnualNet) && hasActiveMonths
           ? bufferedAnnualNet / activeMonths
           : null;
-
+  
         row.columns.push({
           classesPerWeek: column.classesPerWeek,
           classesPerYear,
@@ -3392,7 +3893,7 @@ export function initCalculator() {
             monthlyNet: bufferedMonthlyNet
           }
         });
-
+  
         latestResults.push({
           variant: 'Base (no buffer)',
           students,
@@ -3401,7 +3902,7 @@ export function initCalculator() {
           priceExVat: roundedBaseExVat,
           priceInclVat: roundedBaseInclVat
         });
-
+  
         latestResults.push({
           variant: `Buffered +${bufferPercent}%`,
           students,
@@ -3411,25 +3912,25 @@ export function initCalculator() {
           priceInclVat: roundedBufferedInclVat
         });
       }
-
+  
       pricingData.push(row);
     }
-
+  
     latestResultsMode = mode;
     return { pricingData, bufferPercent, revenueNeeded, mode, manualNetSummary };
   }
-
+  
   function render() {
     const inputs = getInputs();
     latestInputsSnapshot = cloneInputs(inputs);
     const { pricingData, bufferPercent, revenueNeeded, mode, manualNetSummary } = computeTables(inputs);
-
+  
     if (mode === PRICING_MODE_LESSON && manualNetSummary) {
       if (Number.isFinite(manualNetSummary.annual)) {
         writeDesiredIncomeNet('year', Math.max(manualNetSummary.annual, 0));
         inputs.targetNet = manualNetSummary.annual;
       }
-
+  
       if (Number.isFinite(manualNetSummary.weekly)) {
         writeDesiredIncomeNet('week', Math.max(manualNetSummary.weekly, 0));
         inputs.targetNetPerWeek = manualNetSummary.weekly;
@@ -3437,7 +3938,7 @@ export function initCalculator() {
         writeDesiredIncomeNet('week', null);
         inputs.targetNetPerWeek = null;
       }
-
+  
       if (Number.isFinite(manualNetSummary.monthly)) {
         writeDesiredIncomeNet('month', Math.max(manualNetSummary.monthly, 0));
         inputs.targetNetPerMonth = manualNetSummary.monthly;
@@ -3445,7 +3946,7 @@ export function initCalculator() {
         writeDesiredIncomeNet('month', null);
         inputs.targetNetPerMonth = null;
       }
-
+  
       if (Number.isFinite(manualNetSummary.averageWeekly)) {
         writeDesiredIncomeNet('avgWeek', Math.max(manualNetSummary.averageWeekly, 0));
         inputs.targetNetAveragePerWeek = manualNetSummary.averageWeekly;
@@ -3453,7 +3954,7 @@ export function initCalculator() {
         writeDesiredIncomeNet('avgWeek', null);
         inputs.targetNetAveragePerWeek = null;
       }
-
+  
       if (Number.isFinite(manualNetSummary.averageMonthly)) {
         writeDesiredIncomeNet('avgMonth', Math.max(manualNetSummary.averageMonthly, 0));
         inputs.targetNetAveragePerMonth = manualNetSummary.averageMonthly;
@@ -3461,7 +3962,7 @@ export function initCalculator() {
         writeDesiredIncomeNet('avgMonth', null);
         inputs.targetNetAveragePerMonth = null;
       }
-
+  
       refreshDesiredIncomeDisplay(
         {
           year: manualNetSummary.annual,
@@ -3474,14 +3975,14 @@ export function initCalculator() {
       );
       refreshAcceptableIncomeDisplay(inputs.taxRate, { force: true });
     }
-
+  
     latestCurrencySymbol = inputs.currencySymbol;
     latestPricingMode = mode;
     latestBufferPercent = bufferPercent;
     latestPricingData = pricingData.length ? pricingData : null;
-
+  
     closeActiveInfoIcon();
-
+  
     if (!pricingData.length) {
       tablesContainer.innerHTML = `
         <div class="card">
@@ -3514,11 +4015,11 @@ export function initCalculator() {
             workingWeeks: inputs.workingWeeks,
             activeMonths: inputs.activeMonths
           });
-
+  
       tablesContainer.innerHTML = pricingTable + targetsTable;
       scheduleTablesLayoutUpdate();
     }
-
+  
     if (breakdownDialog && !breakdownDialog.hidden) {
       if (!latestPricingData || !refreshBreakdownDialog()) {
         closeBreakdownDialog();
@@ -3527,11 +4028,11 @@ export function initCalculator() {
       activeBreakdownContext = null;
       breakdownTriggerElement = null;
     }
-
+  
     enhanceAllInfoIcons();
     renderAssumptions(inputs, { mode });
   }
-
+  
   function renderAssumptions(inputs, { mode } = {}) {
     const {
       targetNet,
@@ -3563,24 +4064,24 @@ export function initCalculator() {
       lessonPriceMin,
       lessonPriceMax
     } = inputs;
-
+  
     const activeMonthPercentage = activeMonthShare * 100;
     const activeWeeksPercentage = weeksShare * 100;
     const workingWeeksPerCycle = 4 * weeksShare;
-
+  
     const targetPerWeekDisplay = Number.isFinite(targetNetPerWeek)
       ? formatCurrency(currencySymbol, targetNetPerWeek)
-      : '—';
+      : '–';
     const targetPerMonthDisplay = Number.isFinite(targetNetPerMonth)
       ? formatCurrency(currencySymbol, targetNetPerMonth)
-      : '—';
+      : '–';
     const targetAveragePerWeekDisplay = Number.isFinite(targetNetAveragePerWeek)
       ? formatCurrency(currencySymbol, targetNetAveragePerWeek)
-      : '—';
+      : '–';
     const targetAveragePerMonthDisplay = Number.isFinite(targetNetAveragePerMonth)
       ? formatCurrency(currencySymbol, targetNetAveragePerMonth)
-      : '—';
-
+      : '–';
+  
     const manualLessonLine = Number.isFinite(lessonCostInclVat)
       ? `Preferred lesson price (incl. VAT): ${currencySymbol}${formatFixed(lessonCostInclVat, 2)}`
       : 'Preferred lesson price (incl. VAT): not set';
@@ -3590,7 +4091,7 @@ export function initCalculator() {
     const lessonMaxLine = Number.isFinite(lessonPriceMax)
       ? `Maximum preferred lesson price (incl. VAT): ${currencySymbol}${formatFixed(lessonPriceMax, 2)}`
       : 'Maximum preferred lesson price (incl. VAT): not set';
-
+  
     const listItems = [
       `Net income per year: ${formatCurrency(currencySymbol, targetNet)}`,
       `Net income per active week: ${targetPerWeekDisplay}`,
@@ -3618,10 +4119,10 @@ export function initCalculator() {
       `Currency symbol: ${currencySymbol}`,
       `Prices are rounded to whole currency units for display and CSV export.`
     ];
-
+  
     assumptionsList.innerHTML = listItems.map(item => `<li>${item}</li>`).join('');
   }
-
+  
   function downloadCsv() {
     if (!latestResults.length) {
       controls.statusMessage.textContent = 'Add at least one class and student value before exporting CSV.';
@@ -3630,10 +4131,10 @@ export function initCalculator() {
       }, 2500);
       return;
     }
-
+  
     let header;
     let rows;
-
+  
     if (latestResultsMode === PRICING_MODE_LESSON) {
       header = 'Variant,Students,Classes per week,Classes per year,Monthly net income,Annual net income';
       rows = latestResults.map(entry => {
@@ -3659,7 +4160,7 @@ export function initCalculator() {
         entry.priceExVat
       ].join(','));
     }
-
+  
     const csvContent = [header, ...rows].join('\n');
     const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
     const url = URL.createObjectURL(blob);
@@ -3678,7 +4179,7 @@ export function initCalculator() {
       controls.statusMessage.textContent = '';
     }, 2500);
   }
-
+  
   [
     { control: controls.targetNet, basis: 'year' },
     { control: controls.targetNetWeek, basis: 'week' },
@@ -3710,7 +4211,7 @@ export function initCalculator() {
       }
     });
   });
-
+  
   if (controls.acceptableIncomeMin instanceof HTMLInputElement) {
     const updateMin = () => {
       updateAcceptableIncomeFromInput('min');
@@ -3727,7 +4228,7 @@ export function initCalculator() {
       refreshAcceptableIncomeDisplay(getCurrentTaxRate(), { force: true });
     });
   }
-
+  
   if (controls.acceptableIncomeMax instanceof HTMLInputElement) {
     const updateMax = () => {
       updateAcceptableIncomeFromInput('max');
@@ -3744,12 +4245,12 @@ export function initCalculator() {
       refreshAcceptableIncomeDisplay(getCurrentTaxRate(), { force: true });
     });
   }
-
+  
   const acceptableBasisControls = [
     { control: controls.acceptableIncomeBasisMonthly, value: 'monthly' },
     { control: controls.acceptableIncomeBasisAnnual, value: 'annual' }
   ];
-
+  
   acceptableBasisControls.forEach(({ control, value }) => {
     if (!(control instanceof HTMLInputElement)) {
       return;
@@ -3763,7 +4264,7 @@ export function initCalculator() {
       refreshAcceptableIncomeDisplay(getCurrentTaxRate(), { force: true });
     });
   });
-
+  
   if (controls.desiredIncomeToggleDisplay instanceof HTMLButtonElement) {
     controls.desiredIncomeToggleDisplay.addEventListener('click', () => {
       desiredIncomeDisplayMode = desiredIncomeDisplayMode === 'gross' ? 'net' : 'gross';
@@ -3779,7 +4280,7 @@ export function initCalculator() {
       render();
     });
   }
-
+  
   if (controls.desiredIncomeToggleInterpretation instanceof HTMLButtonElement) {
     controls.desiredIncomeToggleInterpretation.addEventListener('click', () => {
       const taxRate = getCurrentTaxRate();
@@ -3835,7 +4336,7 @@ export function initCalculator() {
       }
     });
   }
-
+  
   Object.values(controls).forEach(control => {
     if (!(control instanceof HTMLInputElement)) return;
     if (control === controls.lessonCost) return;
@@ -3845,23 +4346,23 @@ export function initCalculator() {
       render();
     });
   });
-
+  
   if (controls.lessonCost instanceof HTMLInputElement) {
     const handleLessonCostInput = () => {
       render();
     };
-
+  
     controls.lessonCost.addEventListener('input', handleLessonCostInput);
     controls.lessonCost.addEventListener('change', handleLessonCostInput);
   }
-
+  
   window.addEventListener('resize', scheduleTablesLayoutUpdate);
-
+  
   tablesContainer.addEventListener('click', event => {
     const priceButton = event.target instanceof HTMLElement
       ? event.target.closest('button.price-line[data-row][data-column][data-variant]')
       : null;
-
+  
     if (priceButton) {
       const context = {
         rowIndex: Number(priceButton.dataset.row),
@@ -3871,7 +4372,7 @@ export function initCalculator() {
       openBreakdownDialog(context, priceButton);
       return;
     }
-
+  
     const toggleButton = event.target instanceof HTMLElement ? event.target.closest('.price-reveal-toggle') : null;
     if (!toggleButton) {
       return;
@@ -3884,11 +4385,10 @@ export function initCalculator() {
       refreshedToggle.focus();
     }
   });
-
+  
   controls.recalcButton.addEventListener('click', render);
   controls.downloadCsv.addEventListener('click', downloadCsv);
-
+  
   initializePersistence();
   render();
-    
 }
